The H2OClusteringModel object.

This virtual class represents a clustering model built by H2O.

Details

This object has slots for the key, which is a character string that points to the model key existing in the H2O cloud, the data used to build the model (an object of class H2OFrame).

Slots

model_id

A character string specifying the key for the model fit in the H2O cloud's key-value store.

algorithm

A character string specifying the algorithm that was used to fit the model.

parameters

A list containing the parameter settings that were used to fit the model that differ from the defaults.

allparameters

A list containing all parameters used to fit the model.

model

A list containing the characteristics of the model returned by the algorithm.

size

The number of points in each cluster.

totss

Total sum of squared error to grand mean.

withinss

A vector of within-cluster sum of squared error.

tot_withinss

Total within-cluster sum of squared error.

betweenss

Between-cluster sum of squared error.