The H2OConnection class.

This class represents a connection to an H2O cloud.

# S4 method for H2OConnection
show(object)

Arguments

object
an H2OConnection object.

Details

Because H2O is not a master-slave architecture, there is no restriction on which H2O node is used to establish the connection between R (the client) and H2O (the server).

A new H2O connection is established via the h2o.init() function, which takes as parameters the `ip` and `port` of the machine running an instance to connect with. The default behavior is to connect with a local instance of H2O at port 54321, or to boot a new local instance if one is not found at port 54321.

Slots

ip

A character string specifying the IP address of the H2O cloud.

port

A numeric value specifying the port number of the H2O cloud.

proxy

A character specifying the proxy path of the H2O cloud.

https

Set this to TRUE to use https instead of http.

insecure

Set this to TRUE to disable SSL certificate checking.

username

Username to login with.

password

Password to login with.

cookies

Cookies to add to request

context_path

Context path which is appended to H2O server location.

mutable

An H2OConnectionMutableState object to hold the mutable state for the H2O connection.