Apply on H2O Datasets

Method for apply on H2OFrame objects.

apply(X, MARGIN, FUN, ...)

Arguments

X
an H2OFrame object on which apply will operate.

MARGIN
the vector on which the function will be applied over, either 1 for rows or 2 for columns.

FUN
the function to be applied.

…
optional arguments to FUN.

Value

Produces a new H2OFrame of the output of the applied function. The output is stored in H2O so that it can be used in subsequent H2O processes.

See also

apply for the base generic

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.98 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

irisPath <- system.file("extdata", "iris.csv", package="h2o")
iris.hex <- h2o.importFile(path = irisPath, destination_frame = "iris.hex")

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

summary(apply(iris.hex, 2, sum))

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  C1              C2              C3              C4              C5            
#>  Min.   :876.5   Min.   :458.1   Min.   :563.8   Min.   :179.8   Min.   : Inf  
#>  1st Qu.:  NaN   1st Qu.:  NaN   1st Qu.:  NaN   1st Qu.:  NaN   1st Qu.: NaN  
#>  Median :  NaN   Median :  NaN   Median :  NaN   Median :  NaN   Median : NaN  
#>  Mean   :876.5   Mean   :458.1   Mean   :563.8   Mean   :179.8   Mean   :   0  
#>  3rd Qu.:  NaN   3rd Qu.:  NaN   3rd Qu.:  NaN   3rd Qu.:  NaN   3rd Qu.: NaN  
#>  Max.   :876.5   Max.   :458.1   Max.   :563.8   Max.   :179.8   Max.   :-Inf  
#>                                                                  NA's   :1