Converts parsed H2O data into an R data frame

Downloads the H2O data and then scans it in to an R data frame.

# S3 method for H2OFrame
as.data.frame(x, ...)

Arguments

x
An H2OFrame object.

...
Further arguments to be passed down from other methods.

Details

Method as.data.frame.H2OFrame will use fread if data.table package is installed in required version.

See also

use.package

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.98 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

as.data.frame(prostate.hex)

#>      ID CAPSULE AGE RACE DPROS DCAPS    PSA   VOL GLEASON
#> 1     1       0  65    1     2     1   1.40  0.00       6
#> 2     2       0  72    1     3     2   6.70  0.00       7
#> 3     3       0  70    1     1     2   4.90  0.00       6
#> 4     4       0  76    2     2     1  51.20 20.00       7
#> 5     5       0  69    1     1     1  12.30 55.90       6
#> 6     6       1  71    1     3     2   3.30  0.00       8
#> 7     7       0  68    2     4     2  31.90  0.00       7
#> 8     8       0  61    2     4     2  66.70 27.20       7
#> 9     9       0  69    1     1     1   3.90 24.00       7
#> 10   10       0  68    2     1     2  13.00  0.00       6
#> 11   11       1  68    2     4     2   4.00  0.00       7
#> 12   12       1  72    1     2     2  21.20  0.00       7
#> 13   13       1  72    1     4     2  22.70  0.00       9
#> 14   14       1  65    1     4     2  39.00  0.00       7
#> 15   15       0  75    1     1     1   7.50  0.00       5
#> 16   16       0  73    1     2     1   2.60  0.00       5
#> 17   17       0  75    2     1     1   2.50  0.00       5
#> 18   18       0  70    1     2     1   2.60 11.80       5
#> 19   19       0  54    1     1     2   2.80  0.00       6
#> 20   20       1  67    2     3     2   8.60 25.50       7
#> 21   21       1  58    1     2     1   3.10  0.00       7
#> 22   22       1  70    0     4     1  67.10  0.00       7
#> 23   23       1  74    1     3     1  12.70 27.50       7
#> 24   24       0  73    1     1     1  12.30 47.30       7
#> 25   25       1  77    1     1     1  61.10 58.00       7
#> 26   26       0  77    1     1     1   8.80  0.00       5
#> 27   27       0  67    1     2     1   2.80 25.60       7
#> 28   28       0  73    1     3     1   2.90 14.10       5
#> 29   29       0  64    1     3     1   5.60 34.10       5
#> 30   30       0  58    1     4     1   5.60 22.80       6
#> 31   31       1  54    1     3     1   8.40 18.30       6
#> 32   32       0  72    1     2     1   6.50 22.50       7
#> 33   33       0  77    1     4     1  11.00  0.00       8
#> 34   34       1  60    1     3     2   9.50  0.00       7
#> 35   35       0  65    1     2     1  11.10 17.70       6
#> 36   36       0  71    1     2     1   7.30 20.00       7
#> 37   37       0  54    1     2     1   1.00  0.00       6
#> 38   38       1  78    1     1     2  27.20  0.00       8
#> 39   39       1  63    1     2     1  35.10 18.70       7
#> 40   40       1  73    1     3     1   4.50 26.40       7
#> 41   41       1  66    1     3     1   7.90 20.80       7
#> 42   42       0  71    1     1     1   2.00  0.00       6
#> 43   43       1  71    1     2     1   7.50  0.00       6
#> 44   44       0  72    1     2     1   5.30  0.00       7
#> 45   45       1  65    2     3     1  83.70 32.00       9
#> 46   46       1  75    0     3     2  33.00 15.20       8
#> 47   47       0  69    2     2     1   6.70 59.80       6
#> 48   48       0  68    1     2     1  12.30 16.30       8
#> 49   49       0  70    1     2     1   0.40 17.10       5
#> 50   50       0  67    1     2     1   5.80 14.50       6
#> 51   51       0  76    1     1     1   7.60  3.70       6
#> 52   52       1  72    1     4     2 124.00 38.60       8
#> 53   53       0  66    1     3     1   8.80 39.90       6
#> 54   54       1  61    1     3     1  67.60  0.00       7
#> 55   55       0  70    1     3     1  13.90 13.00       7
#> 56   56       1  57    1     3     1   7.40 18.30       7
#> 57   57       0  74    1     2     1  23.10  5.90       7
#> 58   58       1  70    1     3     1  19.30  0.00       7
#> 59   59       0  60    1     3     1   3.00 16.50       6
#> 60   60       0  58    1     3     1   3.70 29.90       6
#> 61   61       0  59    1     2     1   0.70 96.00       5
#> 62   62       1  71    1     2     1   6.00 31.00       6
#> 63   63       0  76    1     2     1   9.50 14.40       7
#> 64   64       0  66    1     3     1   2.60  0.00       7
#> 65   65       1  59    1     4     1  30.70  0.00       7
#> 66   66       1  65    1     2     1  10.80  0.00       7
#> 67   67       0  54    1     2     1  10.50  0.00       6
#> 68   68       0  78    1     1     1   6.50  0.00       7
#> 69   69       0  65    1     2     1   1.30  6.80       5
#> 70   70       1  68    2     3     1   9.60 32.00       6
#> 71   71       0  68    1     2     1   0.30  0.00       6
#> 72   72       0  71    1     3     1   8.30 17.50       5
#> 73   73       0  60    1     3     1   3.20  0.00       7
#> 74   74       0  65    1     3     1   6.90 23.30       5
#> 75   75       0  68    1     1     1  11.00  0.00       7
#> 76   76       0  54    2     1     1  64.30  0.00       7
#> 77   77       0  73    1     3     1   1.60 17.10       6
#> 78   78       1  62    1     2     1   1.90  0.00       6
#> 79   79       0  60    1     2     1   7.90  0.00       5
#> 80   80       1  66    1     1     1  25.70 39.10       9
#> 81   81       0  76    1     3     1   4.90  0.00       6
#> 82   82       0  62    1     1     1  22.10  0.00       7
#> 83   83       0  74    1     2     1  31.50  0.00       7
#> 84   84       0  75    1     2     1  11.00 35.00       7
#> 85   85       0  75    1     1     1   9.90 15.40       7
#> 86   86       1  75    1     3     1   3.70  0.00       6
#> 87   87       1  68    1     4     2  51.30 47.00       9
#> 88   88       1  71    1     2     2  89.00 24.00       8
#> 89   89       0  68    1     4     1  17.10 35.00       9
#> 90   90       0  70    1     3     1  12.30 10.30       7
#> 91   91       0  68    1     1     1   4.40 39.00       6
#> 92   92       0  66    1     1     1   8.00  0.00       5
#> 93   93       1  70    1     3     1  15.00  0.00       7
#> 94   94       1  65    1     3     2  35.80 29.00       9
#> 95   95       0  71    1     1     1  13.40 44.20       7
#> 96   96       1  75    1     1     1  16.00 18.70       7
#> 97   97       0  70    1     3     1  11.20  0.00       7
#> 98   98       0  58    1     1     1   7.00  0.00       6
#> 99   99       0  64    1     1     1  29.10  0.00       6
#> 100 100       1  66    1     3     1   9.50 28.10       7
#> 101 101       0  64    1     2     1   6.10  0.00       6
#> 102 102       1  72    1     3     1   6.30 34.00       7
#> 103 103       0  62    1     1     1   2.80 44.00       6
#> 104 104       0  75    1     3     1  25.70 87.60       5
#> 105 105       1  68    1     2     1   5.70  0.00       7
#> 106 106       0  56    1     3     2   2.70 37.00       6
#> 107 107       1  69    1     1     1   6.00  0.00       7
#> 108 108       0  67    1     3     1  40.40  0.00       7
#> 109 109       0  66    1     3     1  13.20 23.60       6
#> 110 110       1  69    1     3     1  15.20  0.00       7
#> 111 111       0  74    1     1     1   8.80  0.00       7
#>  [ reached getOption("max.print") -- omitted 269 rows ]