Create H2OFrame

Import R object to the H2O cloud.

as.h2o(x, destination_frame = "", ...)

# S3 method for default
as.h2o(x, destination_frame = "", ...)

# S3 method for H2OFrame
as.h2o(x, destination_frame = "", ...)

# S3 method for data.frame
as.h2o(x, destination_frame = "", ...)

# S3 method for Matrix
as.h2o(x, destination_frame = "", ...)

Arguments

x
An R object.

destination_frame
A string with the desired name for the H2OFrame.

…
arguments passed to method arguments.

Details

Method as.h2o.data.frame will use fwrite if data.table package is installed in required version.

To speedup execution time for large sparse matrices, use h2o datatable. Make sure you have installed and imported data.table and slam packages. Turn on h2o datatable by options("h2o.use.data.table"=TRUE)

References

http://blog.h2o.ai/2016/04/fast-csv-writing-for-r/

See also

use.package

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

hi <- as.h2o(iris)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

he <- as.h2o(euro)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

hl <- as.h2o(letters)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

hm <- as.h2o(state.x77)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

hh <- as.h2o(hi)

#> Error in as.h2o(hi): object 'hi' not found

stopifnot(is.h2o(hi), dim(hi)==dim(iris),
          is.h2o(he), dim(he)==c(length(euro),1L),
          is.h2o(hl), dim(hl)==c(length(letters),1L),
          is.h2o(hm), dim(hm)==dim(state.x77),
          is.h2o(hh), dim(hh)==dim(hi))

#> Error in inherits(x, "H2OFrame"): object 'hi' not found

if (requireNamespace("Matrix", quietly=TRUE)) {
  data <- rep(0, 100)
  data[(1:10)^2] <- 1:10 * pi
  m <- matrix(data, ncol = 20, byrow = TRUE)
  m <- Matrix::Matrix(m, sparse = TRUE)
  hs <- as.h2o(m)
  stopifnot(is.h2o(hs), dim(hs)==dim(m))
}

#> Warning: as.h2o can be slow for large sparse matrices.  Install packages data.table and slam to speed up as.h2o.

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?