Access H2O Confusion Matrices

Retrieve either a single or many confusion matrices from H2O objects.

h2o.confusionMatrix(object, ...)

# S4 method for H2OModel
h2o.confusionMatrix(object, newdata, valid = FALSE, ...)

# S4 method for H2OModelMetrics
h2o.confusionMatrix(object, thresholds = NULL,
  metrics = NULL)

Arguments

object
Either an H2OModel object or an H2OModelMetrics object.

...
Extra arguments for extracting train or valid confusion matrices.

newdata
An H2OFrame object that can be scored on. Requires a valid response column.

valid
Retrieve the validation metric.

thresholds
(Optional) A value or a list of valid values between 0.0 and 1.0. This value is only used in the case of H2OBinomialMetrics objects.

metrics
(Optional) A metric or a list of valid metrics ("min_per_class_accuracy", "absolute_mcc", "tnr", "fnr", "fpr", "tpr", "precision", "accuracy", "f0point5", "f2", "f1"). This value is only used in the case of H2OBinomialMetrics objects.

Value

Calling this function on H2OModel objects returns a confusion matrix corresponding to the predict function. If used on an H2OBinomialMetrics object, returns a list of matrices corresponding to the number of thresholds specified.

Details

The H2OModelMetrics version of this function will only take H2OBinomialMetrics or H2OMultinomialMetrics objects. If no threshold is specified, all possible thresholds are selected.

See also

predict for generating prediction frames, h2o.performance for creating H2OModelMetrics.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
hex <- h2o.uploadFile(prosPath)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

hex[,2] <- as.factor(hex[,2])

#> Error in is.H2OFrame(x): object 'hex' not found

model <- h2o.gbm(x = 3:9, y = 2, training_frame = hex, distribution = "bernoulli")

#> Error in is.H2OFrame(training_frame): object 'hex' not found

h2o.confusionMatrix(model, hex)

#> Error in h2o.confusionMatrix(model, hex): object 'model' not found

# Generating a ModelMetrics object
perf <- h2o.performance(model, hex)

#> Error in is(model, "H2OModel"): object 'model' not found

h2o.confusionMatrix(perf)

#> Error in h2o.confusionMatrix(perf): object 'perf' not found