Connect to a running H2O instance.

Connect to a running H2O instance.

h2o.connect(ip = "localhost", port = 54321,
  strict_version_check = TRUE, proxy = NA_character_, https = FALSE,
  insecure = FALSE, username = NA_character_,
  password = NA_character_, cookies = NA_character_,
  context_path = NA_character_, config = NULL)

Arguments

ip
Object of class character representing the IP address of the server where H2O is running.

port
Object of class numeric representing the port number of the H2O server.

strict_version_check
(Optional) Setting this to FALSE is unsupported and should only be done when advised by technical support.

proxy
(Optional) A character string specifying the proxy path.

https
(Optional) Set this to TRUE to use https instead of http.

insecure
(Optional) Set this to TRUE to disable SSL certificate checking.

username
(Optional) Username to login with.

password
(Optional) Password to login with.

cookies
(Optional) Vector(or list) of cookies to add to request.

context_path
(Optional) The last part of connection URL: http://<ip>:<port>/<context_path>

config
(Optional) A list describing connection parameters. Using config makes h2o.connect ignore other parameters and collect named list members instead (see examples).

Value

an instance of H2OConnection object representing a connection to the running H2O instance.

Examples

# NOT RUN {
library(h2o)
# Try to connect to a H2O instance running at http://localhost:54321/cluster_X
# If not found, start a local H2O instance from R with the default settings.
#h2o.connect(ip = "localhost", port = 54321, context_path = "cluster_X")
# Or
#config = list(ip = "localhost", port = 54321, context_path = "cluster_X")
#h2o.connect(config = config)

# Skip strict version check during connecting to the instance
#h2o.connect(config = c(strict_version_check = FALSE, config))
# }