Cut H2O Numeric Data to Factor

Divides the range of the H2O data into intervals and codes the values according to which interval they fall in. The leftmost interval corresponds to the level one, the next is level two, etc.

h2o.cut(x, breaks, labels = NULL, include.lowest = FALSE,
  right = TRUE, dig.lab = 3, ...)

# S3 method for H2OFrame
cut(x, breaks, labels = NULL,
  include.lowest = FALSE, right = TRUE, dig.lab = 3, ...)

Arguments

x
An H2OFrame object with a single numeric column.

breaks
A numeric vector of two or more unique cut points.

labels
Labels for the levels of the resulting category. By default, labels are constructed sing "(a,b]" interval notation.

include.lowest
Logical, indicationg if an 'x[i]' equal to the lowest (or highest, for right = FALSE 'breaks' value should be included

right
/codeLogical, indicating if the intervals should be closed on the right (opened on the left) or vice versa.

dig.lab
Integer which is used when labels are not given, determines the number of digits used in formatting the break numbers.

...
Further arguments passed to or from other methods.

Value

Returns an H2OFrame object containing the factored data with intervals as levels.

Examples

# NOT RUN {
library(h2o)
h2o.init()
iris_hf <- as.h2o(iris)
summary(iris_hf)

# Cut sepal length column into intervals determined by min/max/quantiles
sepal_len_cut <- cut(iris_hf$Sepal.Length, c(4.2, 4.8, 5.8, 6, 8))
head(sepal_len_cut)
summary(sepal_len_cut)
# }