Compute DCT of an H2OFrame

Compute the Discrete Cosine Transform of every row in the H2OFrame

h2o.dct(data, destination_frame, dimensions, inverse = FALSE)

Arguments

data
An H2OFrame object representing the dataset to transform

destination_frame
A frame ID for the result

dimensions
An array containing the 3 integer values for height, width, depth of each sample. The product of HxWxD must total up to less than the number of columns. For 1D, use c(L,1,1), for 2D, use C(N,M,1).

inverse
Whether to perform the inverse transform

Value

Returns an H2OFrame object.

Examples

  library(h2o)
  h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

  df <- h2o.createFrame(rows = 1000, cols = 8*16*24,
                        categorical_fraction = 0, integer_fraction = 0, missing_fraction = 0)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

  df1 <- h2o.dct(data=df, dimensions=c(8*16*24,1,1))

#> Error in chk.H2OFrame(x): must be an H2OFrame

  df2 <- h2o.dct(data=df1,dimensions=c(8*16*24,1,1),inverse=TRUE)

#> Error in is.H2OFrame(fr): object 'df1' not found

  max(abs(df1-df2))

#> Error in eval(expr, envir, enclos): object 'df1' not found

  df1 <- h2o.dct(data=df, dimensions=c(8*16,24,1))

#> Error in chk.H2OFrame(x): must be an H2OFrame

  df2 <- h2o.dct(data=df1,dimensions=c(8*16,24,1),inverse=TRUE)

#> Error in is.H2OFrame(fr): object 'df1' not found

  max(abs(df1-df2))

#> Error in eval(expr, envir, enclos): object 'df1' not found

  df1 <- h2o.dct(data=df, dimensions=c(8,16,24))

#> Error in chk.H2OFrame(x): must be an H2OFrame

  df2 <- h2o.dct(data=df1,dimensions=c(8,16,24),inverse=TRUE)

#> Error in is.H2OFrame(fr): object 'df1' not found

  max(abs(df1-df2))

#> Error in eval(expr, envir, enclos): object 'df1' not found