Compute a pairwise distance measure between all rows of two numeric H2OFrames.

Compute a pairwise distance measure between all rows of two numeric H2OFrames.

h2o.distance(x, y, measure)

Arguments

x
An H2OFrame object (large, references).

y
An H2OFrame object (small, queries).

measure
An optional string indicating what distance measure to use. Must be one of: "l1" - Absolute distance (L1-norm, >=0) "l2" - Euclidean distance (L2-norm, >=0) "cosine" - Cosine similarity (-1...1) "cosine_sq" - Squared Cosine similarity (0...1)

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.83 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

h2o.distance(prostate.hex[11:30,], prostate.hex[1:10,], "cosine")

#>          C1        C2        C3        C4        C5        C6        C7
#> 1 0.9883124 0.9904351 0.9918635 0.8366993 0.7743389 0.9968006 0.9285223
#> 2 0.9556401 0.9730989 0.9703707 0.9258961 0.7759796 0.9691321 0.9860979
#> 3 0.9485704 0.9677477 0.9638187 0.9289561 0.7725520 0.9640587 0.9879995
#> 4 0.8578121 0.8922987 0.8838388 0.9657851 0.7283872 0.8794001 0.9908468
#> 5 0.9802759 0.9848838 0.9875305 0.8522941 0.7752535 0.9906916 0.9382629
#> 6 0.9797437 0.9803078 0.9842620 0.8193900 0.7679497 0.9900876 0.9144861
#>          C8        C9       C10
#> 1 0.6918676 0.9444152 0.9905390
#> 2 0.8209062 0.9212774 0.9949971
#> 3 0.8292782 0.9162058 0.9917088
#> 4 0.9150974 0.8391193 0.9394290
#> 5 0.7149201 0.9424985 0.9943186
#> 6 0.6732078 0.9417636 0.9856776
#> 
#> [20 rows x 10 columns]