Searches for matches to argument `pattern` within each element of a string column.

This function has similar semantics as R's native grep function and it supports a subset of its parameters. Default behavior is to return indices of the elements matching the pattern. Parameter `output.logical` can be used to return a logical vector indicating if the element matches the pattern (1) or not (0).

h2o.grep(pattern, x, ignore.case = FALSE, invert = FALSE,
  output.logical = FALSE)

Arguments

pattern
A character string containing a regular expression.

x
An H2O frame that wraps a single string column.

ignore.case
If TRUE case is ignored during matching.

invert
Identify elements that do not match the pattern.

output.logical
If TRUE returns logical vector of indicators instead of list of matching positions

Value

H2OFrame holding the matching positions or a logical vector if `output.logical` is enabled.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

addresses <- as.h2o(c("2307", "Leghorn St", "Mountain View", "CA", "94043"))

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

zip.codes <- addresses[h2o.grep("[0-9]{5}", addresses, output.logical = TRUE),]

#> Error in eval(expr, envir, enclos): object 'addresses' not found