Return the Head or Tail of an H2O Dataset.

Returns the first or last rows of an H2OFrame object.

h2o.head(x, n = 6L, ...)

# S3 method for H2OFrame
head(x, n = 6L, ...)

h2o.tail(x, n = 6L, ...)

# S3 method for H2OFrame
tail(x, n = 6L, ...)

Arguments

x
An H2OFrame object.

n
(Optional) A single integer. If positive, number of rows in x to return. If negative, all but the n first/last number of rows in x.

...
Ignored.

Value

An H2OFrame containing the first or last n rows of an H2OFrame object.

Examples

library(h2o)
h2o.init(ip <- "localhost", port = 54321, startH2O = TRUE)

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

ausPath <- system.file("extdata", "australia.csv", package="h2o")
australia.hex <- h2o.uploadFile(path = ausPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

head(australia.hex, 10)

#>    premax salmax minairtemp maxairtemp maxsst maxsoilmoist Max_czcs runoffnew
#> 1  350.00   3448      287.0      301.5 302817         6.39     0.51       167
#> 2  350.00   3487      290.0      301.2 302280         0.00     0.80         0
#> 3  450.00   3483      289.0      301.0 303326         1.00     1.25         0
#> 4  378.57   3467      286.0      302.0 303122         7.00     1.65         0
#> 5  350.00   3459      288.0      301.0 302890         5.00     1.53       300
#> 6  375.00   3472      287.0      302.0 302767         9.00     0.75         0
#> 7  435.71   3481      288.1      302.3 302750         0.00     0.82         0
#> 8  435.71   3483      288.9      302.4 302733         0.00     0.82         0
#> 9  450.00   3480      289.0      302.0 302775         9.00     0.97         0
#> 10 350.00   3467      289.9      301.4 302847         0.00     0.68         0

tail(australia.hex, 10)

#>    premax salmax minairtemp maxairtemp maxsst maxsoilmoist Max_czcs runoffnew
#> 1  250.00   3462      273.0      285.0 287680        16.00     2.77      2400
#> 2   75.00   3446      273.0      286.0 286833         9.00     2.35      1200
#> 3   85.71   3442      272.6      286.5 286703         8.25     1.28       764
#> 4  216.67   3465      273.3      285.3 286957         9.20     2.90      1200
#> 5  194.44   3456      273.7      285.5 286440        10.06     3.46      1154
#> 6  150.00   3456      274.0      286.0 286420        14.00     4.01      1200
#> 7  150.00   3448      275.0      286.0 286082        10.00     1.70       600
#> 8  112.50   3444      273.8      286.0 286090         9.63     1.05       960
#> 9  183.33   3456      273.7      285.7 285937         0.00     2.39         0
#> 10 150.00   3453      274.3      285.9 285697         8.22     1.56       600