H2O Apply Conditional Statement

Applies conditional statements to numeric vectors in H2O parsed data objects when the data are numeric.

h2o.ifelse(test, yes, no)

ifelse(test, yes, no)

Arguments

test
A logical description of the condition to be met (>, <, =, etc...)

yes
The value to return if the condition is TRUE.

no
The value to return if the condition is FALSE.

Value

Returns a vector of new values matching the conditions stated in the ifelse call.

Details

Both numeric and categorical values can be tested. However when returning a yes and no condition both conditions must be either both categorical or numeric.

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

ausPath <- system.file("extdata", "australia.csv", package="h2o")
australia.hex <- h2o.importFile(path = ausPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

australia.hex[,9] <- ifelse(australia.hex[,3] < 279.9, 1, 0)
summary(australia.hex)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  premax           salmax         minairtemp      maxairtemp     
#>  Min.   : 18.00   Min.   :3441   Min.   :272.6   Min.   :285.0  
#>  1st Qu.: 74.59   1st Qu.:3490   1st Qu.:277.0   1st Qu.:292.0  
#>  Median :149.76   Median :3533   Median :278.8   Median :299.9  
#>  Mean   :161.46   Mean   :3529   Mean   :279.9   Mean   :297.5  
#>  3rd Qu.:249.98   3rd Qu.:3558   3rd Qu.:282.0   3rd Qu.:302.3  
#>  Max.   :450.00   Max.   :3650   Max.   :290.0   Max.   :310.0  
#>  maxsst           maxsoilmoist     Max_czcs          runoffnew        
#>  Min.   :285697   Min.   : 0.000   Min.   : 0.1600   Min.   :   0.00  
#>  1st Qu.:290485   1st Qu.: 0.000   1st Qu.: 0.6308   1st Qu.:   0.00  
#>  Median :293635   Median : 4.000   Median : 1.0120   Median :  16.81  
#>  Mean   :295676   Mean   : 5.117   Mean   : 1.3694   Mean   : 232.22  
#>  3rd Qu.:301934   3rd Qu.: 8.992   3rd Qu.: 1.7014   3rd Qu.: 297.72  
#>  Max.   :303697   Max.   :16.000   Max.   :11.3700   Max.   :2400.00  
#>  C9              
#>  Min.   :0.0000  
#>  1st Qu.:0.0000  
#>  Median :1.0000  
#>  Mean   :0.5737  
#>  3rd Qu.:1.0000  
#>  Max.   :1.0000