Insert Missing Values into an H2OFrame

Randomly replaces a user-specified fraction of entries in an H2O dataset with missing values.

h2o.insertMissingValues(data, fraction = 0.1, seed = -1)

Arguments

data
An H2OFrame object representing the dataset.

fraction
A number between 0 and 1 indicating the fraction of entries to replace with missing.

seed
A random number used to select which entries to replace with missing values. Default of seed = -1 will automatically generate a seed in H2O.

Value

Returns an H2OFrame object.

WARNING

This will modify the original dataset. Unless this is intended, this function should only be called on a subset of the original.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

irisPath <- system.file("extdata", "iris.csv", package = "h2o")
iris.hex <- h2o.importFile(path = irisPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

summary(iris.hex)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  C1              C2              C3              C4              
#>  Min.   :4.300   Min.   :2.000   Min.   :1.000   Min.   :0.1000  
#>  1st Qu.:5.099   1st Qu.:2.799   1st Qu.:1.596   1st Qu.:0.2992  
#>  Median :5.798   Median :2.998   Median :4.348   Median :1.3000  
#>  Mean   :5.843   Mean   :3.054   Mean   :3.759   Mean   :1.1987  
#>  3rd Qu.:6.399   3rd Qu.:3.298   3rd Qu.:5.095   3rd Qu.:1.7992  
#>  Max.   :7.900   Max.   :4.400   Max.   :6.900   Max.   :2.5000  
#>  C5                 
#>  Iris-setosa    :50 
#>  Iris-versicolor:50 
#>  Iris-virginica :50 
#>                     
#>                     
#>                     

irismiss.hex <- h2o.insertMissingValues(iris.hex, fraction = 0.25)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

head(irismiss.hex)

#>    C1  C2  C3  C4          C5
#> 1 NaN 3.5 1.4 NaN Iris-setosa
#> 2 4.9 3.0 1.4 0.2        <NA>
#> 3 NaN 3.2 1.3 0.2 Iris-setosa
#> 4 NaN 3.1 1.5 NaN Iris-setosa
#> 5 5.0 3.6 1.4 NaN Iris-setosa
#> 6 NaN NaN 1.7 NaN Iris-setosa

summary(irismiss.hex)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  C1              C2              C3              C4              
#>  Min.   :4.300   Min.   :2.000   Min.   :1.000   Min.   :0.1000  
#>  1st Qu.:5.099   1st Qu.:2.799   1st Qu.:1.496   1st Qu.:0.2992  
#>  Median :5.697   Median :2.999   Median :4.198   Median :1.3984  
#>  Mean   :5.792   Mean   :3.041   Mean   :3.766   Mean   :1.2860  
#>  3rd Qu.:6.398   3rd Qu.:3.298   3rd Qu.:5.098   3rd Qu.:1.9000  
#>  Max.   :7.700   Max.   :4.200   Max.   :6.700   Max.   :2.5000  
#>  NA's   :45      NA's   :38      NA's   :37      NA's   :36      
#>  C5                 
#>  Iris-setosa    :41 
#>  Iris-versicolor:37 
#>  Iris-virginica :33 
#>  NA             :39 
#>                     
#>                     
#>