Load H2O Model from HDFS or Local Disk

Load a saved H2O model from disk. (Note that ensemble binary models can now be loaded using this method.)

h2o.loadModel(path)

Arguments

path
The path of the H2O Model to be imported.

Value

Returns a H2OModel object of the class corresponding to the type of model loaded.

See also

Examples

# NOT RUN {
# library(h2o)
# h2o.init()
# prostate_path = system.file("extdata", "prostate.csv", package = "h2o")
# prostate = h2o.importFile(path = prostate_path)
# prostate_glm = h2o.glm(y = "CAPSULE", x = c("AGE","RACE","PSA","DCAPS"),
#   training_frame = prostate, family = "binomial", alpha = 0.5)
# glmmodel_path = h2o.saveModel(prostate_glm, dir = "/Users/UserName/Desktop")
# glmmodel_load = h2o.loadModel(glmmodel_path)
# }