H2O Model Metric Accessor Functions

A series of functions that retrieve model metric details.

h2o.metric(object, thresholds, metric)

h2o.F0point5(object, thresholds)

h2o.F1(object, thresholds)

h2o.F2(object, thresholds)

h2o.accuracy(object, thresholds)

h2o.error(object, thresholds)

h2o.maxPerClassError(object, thresholds)

h2o.mean_per_class_accuracy(object, thresholds)

h2o.mcc(object, thresholds)

h2o.precision(object, thresholds)

h2o.tpr(object, thresholds)

h2o.fpr(object, thresholds)

h2o.fnr(object, thresholds)

h2o.tnr(object, thresholds)

h2o.recall(object, thresholds)

h2o.sensitivity(object, thresholds)

h2o.fallout(object, thresholds)

h2o.missrate(object, thresholds)

h2o.specificity(object, thresholds)

Arguments

object
An H2OModelMetrics object of the correct type.

thresholds
(Optional) A value or a list of values between 0.0 and 1.0.

metric
(Optional) A specified paramter to retrieve.

Value

Returns either a single value, or a list of values.

Details

Many of these functions have an optional thresholds parameter. Currently only increments of 0.1 are allowed. If not specified, the functions will return all possible values. Otherwise, the function will return the value for the indicated threshold.

Currently, the these functions are only supported by H2OBinomialMetrics objects.

See also

h2o.auc for AUC, h2o.giniCoef for the GINI coefficient, and h2o.mse for MSE. See h2o.performance for creating H2OModelMetrics objects.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
hex <- h2o.uploadFile(prosPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

hex[,2] <- as.factor(hex[,2])
model <- h2o.gbm(x = 3:9, y = 2, training_frame = hex, distribution = "bernoulli")

#>   |                                                                              |                                                                      |   0%  |                                                                              |=========================                                             |  36%  |                                                                              |======================================================================| 100%

perf <- h2o.performance(model, hex)
h2o.F1(perf)

#>   threshold         f1
#> 1 0.9824807 0.01298701
#> 2 0.9676546 0.02580645
#> 3 0.9671792 0.03846154
#> 4 0.9649031 0.05095541
#> 5 0.9621187 0.06329114
#> 
#> ---
#>       threshold        f1
#> 375 0.015166897 0.5795455
#> 376 0.014322204 0.5784499
#> 377 0.013740748 0.5773585
#> 378 0.013256314 0.5762712
#> 379 0.012183072 0.5751880
#> 380 0.008968162 0.5741088