Quantiles of H2O Frames.

Obtain and display quantiles for H2O parsed data.

h2o.quantile(x, probs = c(0.001, 0.01, 0.1, 0.25, 0.333, 0.5, 0.667,
  0.75, 0.9, 0.99, 0.999), combine_method = c("interpolate", "average",
  "avg", "low", "high"), weights_column = NULL, ...)

# S3 method for H2OFrame
quantile(x, probs = c(0.001, 0.01, 0.1, 0.25, 0.333,
  0.5, 0.667, 0.75, 0.9, 0.99, 0.999), combine_method = c("interpolate",
  "average", "avg", "low", "high"), weights_column = NULL, ...)

Arguments

x
An H2OFrame object with a single numeric column.

probs
Numeric vector of probabilities with values in [0,1].

combine_method
How to combine quantiles for even sample sizes. Default is to do linear interpolation. E.g., If method is "lo", then it will take the lo value of the quantile. Abbreviations for average, low, and high are acceptable (avg, lo, hi).

weights_column
(Optional) String name of the observation weights column in x or an H2OFrame object with a single numeric column of observation weights.

...
Further arguments passed to or from other methods.

Value

A vector describing the percentiles at the given cutoffs for the H2OFrame object.

Details

quantile.H2OFrame, a method for the quantile generic. Obtain and return quantiles for an H2OFrame object.

Examples

# NOT RUN {
# Request quantiles for an H2O parsed data set:
library(h2o)
h2o.init()
prostate_path <- system.file("extdata", "prostate.csv", package = "h2o")
prostate <- h2o.uploadFile(path = prostate_path)
# Request quantiles for a subset of columns in an H2O parsed data set
quantile(prostate[,3])
for(i in 1:ncol(prostate))
   quantile(prostate[, i])
# }