Save an H2O Model Object to Disk

Save an H2OModel to disk. (Note that ensemble binary models can be saved.)

h2o.saveModel(object, path = "", force = FALSE)

Arguments

object
an H2OModel object.

path
string indicating the directory the model will be written to.

force
logical, indicates how to deal with files that already exist.

Details

In the case of existing files force = TRUE will overwrite the file. Otherwise, the operation will fail.

See also

Examples

# NOT RUN {
# library(h2o)
# h2o.init()
# prostate <- h2o.importFile(path = paste("https://raw.github.com",
#    "h2oai/h2o-2/master/smalldata/logreg/prostate.csv", sep = "/"))
# prostate_glm <- h2o.glm(y = "CAPSULE", x = c("AGE","RACE","PSA","DCAPS"),
#    training_frame = prostate, family = "binomial", alpha = 0.5)
# h2o.saveModel(object = prostate_glm, path = "/Users/UserName/Desktop", force = TRUE)
# }