Scaling and Centering of an H2OFrame

Centers and/or scales the columns of an H2O dataset.

h2o.scale(x, center = TRUE, scale = TRUE)

# S3 method for H2OFrame
scale(x, center = TRUE, scale = TRUE)

Arguments

x
An H2OFrame object.

center
either a logical value or numeric vector of length equal to the number of columns of x.

scale
either a logical value or numeric vector of length equal to the number of columns of x.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   17.59 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

irisPath <- system.file("extdata", "iris_wheader.csv", package="h2o")
iris.hex <- h2o.uploadFile(path = irisPath, destination_frame = "iris.hex")

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

summary(iris.hex)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  sepal_len       sepal_wid       petal_len       petal_wid       
#>  Min.   :4.300   Min.   :2.000   Min.   :1.000   Min.   :0.1000  
#>  1st Qu.:5.099   1st Qu.:2.799   1st Qu.:1.596   1st Qu.:0.2992  
#>  Median :5.798   Median :2.998   Median :4.348   Median :1.3000  
#>  Mean   :5.843   Mean   :3.054   Mean   :3.759   Mean   :1.1987  
#>  3rd Qu.:6.399   3rd Qu.:3.298   3rd Qu.:5.095   3rd Qu.:1.7992  
#>  Max.   :7.900   Max.   :4.400   Max.   :6.900   Max.   :2.5000  
#>  class              
#>  Iris-setosa    :50 
#>  Iris-versicolor:50 
#>  Iris-virginica :50 
#>                     
#>                     
#>                     

# Scale and center all the numeric columns in iris data set
scale(iris.hex[, 1:4])

#>    sepal_len  sepal_wid petal_len petal_wid
#> 1 -0.8976739  1.0286113 -1.336794 -1.308593
#> 2 -1.1392005 -0.1245404 -1.336794 -1.308593
#> 3 -1.3807271  0.3367203 -1.393470 -1.308593
#> 4 -1.5014904  0.1060900 -1.280118 -1.308593
#> 5 -1.0184372  1.2592416 -1.336794 -1.308593
#> 6 -0.5353840  1.9511326 -1.166767 -1.046525
#> 
#> [150 rows x 4 columns]