Standard Deviation of a column of data.

Obtain the standard deviation of a column of data.

h2o.sd(x, na.rm = FALSE)

sd(x, na.rm = FALSE)

Arguments

x
An H2OFrame object.

na.rm
logical. Should missing values be removed?

See also

h2o.var for variance, and sd for the base R implementation.

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

sd(prostate.hex$AGE)

#> Error in is.H2OFrame(x): object 'prostate.hex' not found