Set Levels of H2O Factor Column

Works on a single categorical vector. New domains must be aligned with the old domains. This call has SIDE EFFECTS and mutates the column in place (change of the levels will also affect all the frames that are referencing this column). If you want to make a copy of the column instead, use parameter in.place = FALSE.

h2o.setLevels(x, levels, in.place = TRUE)

Arguments

x
A single categorical column.

levels
A character vector specifying the new levels. The number of new levels must match the number of old levels.

in.place
Indicates whether new domain will be directly applied to the column (in place change) or if a copy of the column will be created with the given domain levels.

Examples

# NOT RUN {
h2o.init()
iris_hf <- as.h2o(iris)
new.levels <- c("setosa", "versicolor", "caroliniana")
iris_hf$Species <- h2o.setLevels(iris_hf$Species, new.levels, in.place = FALSE)
h2o.levels(iris_hf$Species)
# }