Summarizes the columns of an H2OFrame.

A method for the summary generic. Summarizes the columns of an H2O data frame or subset of columns and rows using vector notation (e.g. dataset[row, col]).

h2o.summary(object, factors = 6L, exact_quantiles = FALSE, ...)

# S3 method for H2OFrame
summary(object, factors, exact_quantiles, ...)

Arguments

object
An H2OFrame object.

factors
The number of factors to return in the summary. Default is the top 6.

exact_quantiles
Compute exact quantiles or use approximation. Default is to use approximation.

...
Further arguments passed to or from other methods.

Value

A table displaying the minimum, 1st quartile, median, mean, 3rd quartile and maximum for each numeric column, and the levels and category counts of the levels in each categorical column.

Details

By default it uses approximated version of quantiles computation, however, user can modify this behavior by setting up exact_quantiles argument to true.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.importFile(path = prosPath)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

summary(prostate.hex)

#> Error in summary(prostate.hex): object 'prostate.hex' not found

summary(prostate.hex$GLEASON)

#> Error in summary(prostate.hex$GLEASON): object 'prostate.hex' not found

summary(prostate.hex[,4:6])

#> Error in summary(prostate.hex[, 4:6]): object 'prostate.hex' not found

summary(prostate.hex, exact_quantiles=TRUE)

#> Error in summary(prostate.hex, exact_quantiles = TRUE): object 'prostate.hex' not found