Cross Tabulation and Table Creation in H2O

Uses the cross-classifying factors to build a table of counts at each combination of factor levels.

h2o.table(x, y = NULL, dense = TRUE)

table.H2OFrame(x, y = NULL, dense = TRUE)

Arguments

x
An H2OFrame object with at most two columns.

y
An H2OFrame similar to x, or NULL.

dense
A logical for dense representation, which lists only non-zero counts, 1 combination per row. Set to FALSE to expand counts across all combinations.

Value

Returns a tabulated H2OFrame object.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   17.59 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath, destination_frame = "prostate.hex")

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

summary(prostate.hex)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  ID               CAPSULE          AGE             RACE           
#>  Min.   :  1.00   Min.   :0.0000   Min.   :43.00   Min.   :0.000  
#>  1st Qu.: 95.75   1st Qu.:0.0000   1st Qu.:62.00   1st Qu.:1.000  
#>  Median :190.50   Median :0.0000   Median :67.00   Median :1.000  
#>  Mean   :190.50   Mean   :0.4026   Mean   :66.04   Mean   :1.087  
#>  3rd Qu.:285.25   3rd Qu.:1.0000   3rd Qu.:71.00   3rd Qu.:1.000  
#>  Max.   :380.00   Max.   :1.0000   Max.   :79.00   Max.   :2.000  
#>  DPROS           DCAPS           PSA               VOL            
#>  Min.   :1.000   Min.   :1.000   Min.   :  0.300   Min.   : 0.00  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:  4.900   1st Qu.: 0.00  
#>  Median :2.000   Median :1.000   Median :  8.664   Median :14.20  
#>  Mean   :2.271   Mean   :1.108   Mean   : 15.409   Mean   :15.81  
#>  3rd Qu.:3.000   3rd Qu.:1.000   3rd Qu.: 17.063   3rd Qu.:26.40  
#>  Max.   :4.000   Max.   :2.000   Max.   :139.700   Max.   :97.60  
#>  GLEASON        
#>  Min.   :0.000  
#>  1st Qu.:6.000  
#>  Median :6.000  
#>  Mean   :6.384  
#>  3rd Qu.:7.000  
#>  Max.   :9.000  

# Counts of the ages of all patients
head(h2o.table(prostate.hex[,3]))

#>   AGE Count
#> 1  43     1
#> 2  47     1
#> 3  50     2
#> 4  51     3
#> 5  52     2
#> 6  53     4

h2o.table(prostate.hex[,3])

#>   AGE Count
#> 1  43     1
#> 2  47     1
#> 3  50     2
#> 4  51     3
#> 5  52     2
#> 6  53     4
#> 
#> [32 rows x 2 columns] 

# Two-way table of ages (rows) and race (cols) of all patients
head(h2o.table(prostate.hex[,c(3,4)]))

#>   AGE RACE Counts
#> 1  43    0      1
#> 2  47    1      1
#> 3  50    1      2
#> 4  51    1      3
#> 5  52    1      2
#> 6  53    1      3

h2o.table(prostate.hex[,c(3,4)])

#>   AGE RACE Counts
#> 1  43    0      1
#> 2  47    1      1
#> 3  50    1      2
#> 4  51    1      3
#> 5  52    1      2
#> 6  53    1      3
#> 
#> [53 rows x 3 columns]