Transform words (or sequences of words) to vectors using a word2vec model.

Transform words (or sequences of words) to vectors using a word2vec model.

h2o.transform(word2vec, words, aggregate_method = c("NONE", "AVERAGE"))

Arguments

word2vec
A word2vec model.

words
An H2OFrame made of a single column containing source words.

aggregate_method
Specifies how to aggregate sequences of words. If method is `NONE` then no aggregation is performed and each input word is mapped to a single word-vector. If method is 'AVERAGE' then input is treated as sequences of words delimited by NA. Each word of a sequences is internally mapped to a vector and vectors belonging to the same sentence are averaged and returned in the result.

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

# Build a dummy word2vec model
data <- as.character(as.h2o(c("a", "b", "a")))

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

w2v.model <- h2o.word2vec(data, sent_sample_rate = 0, min_word_freq = 0, epochs = 1, vec_size = 2)

#> Error in value[[3L]](cond): argument 'training_frame' must be a valid H2OFrame or key

# Transform words to vectors without aggregation
sentences <- as.character(as.h2o(c("b", "c", "a", NA, "b")))

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

h2o.transform(w2v.model, sentences) # -> 5 rows total, 2 rows NA ("c" is not in the vocabulary)

#> Error in is(word2vec, "H2OModel"): object 'w2v.model' not found

# Transform words to vectors and return average vector for each sentence
h2o.transform(w2v.model, sentences, aggregate_method = "AVERAGE") # -> 2 rows

#> Error in is(word2vec, "H2OModel"): object 'w2v.model' not found