Plot an H2O Tabulate Heatmap

Plots the simple co-occurrence based tabulation of X vs Y as a heatmap, where X and Y are two Vecs in a given dataset. This function requires suggested ggplot2 package.

# S3 method for H2OTabulate
plot(x, xlab = x$cols[1], ylab = x$cols[2],
  base_size = 12, ...)

Arguments

x
An H2OTabulate object for which the heatmap plot is desired.

xlab
A title for the x-axis. Defaults to what is specified in the given H2OTabulate object.

ylab
A title for the y-axis. Defaults to what is specified in the given H2OTabulate object.

base_size
Base font size for plot.

...
additional arguments to pass on.

Value

Returns a ggplot2-based heatmap of co-occurance.

See also

Examples

# NOT RUN {
library(h2o)
h2o.init()
df <- as.h2o(iris)
tab <- h2o.tabulate(data = df, x = "Sepal.Length", y = "Petal.Width",
             weights_column = NULL, nbins_x = 10, nbins_y = 10)
plot(tab)
# }