Predict the Leaf Node Assignment on an H2O Model

Obtains leaf node assignment from fitted H2O model objects.

predict_leaf_node_assignment.H2OModel(object, newdata, ...)

h2o.predict_leaf_node_assignment(object, newdata, ...)

Arguments

object
a fitted H2OModel object for which prediction is desired

newdata
An H2OFrame object in which to look for variables with which to predict.

...
additional arguments to pass on.

Value

Returns an H2OFrame object with categorical leaf assignment identifiers for each tree in the model.

Details

For every row in the test set, return a set of factors that identify the leaf placements of the row in all the trees in the model. The order of the rows in the results is the same as the order in which the data was loaded

See also

h2o.gbm and h2o.randomForest for model generation in h2o.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   17.59 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

prostate.hex$CAPSULE <- as.factor(prostate.hex$CAPSULE)
prostate.gbm <- h2o.gbm(3:9, "CAPSULE", prostate.hex)

#>   |                                                                              |                                                                      |   0%  |                                                                              |====================================                                  |  52%  |                                                                              |======================================================================| 100%

h2o.predict(prostate.gbm, prostate.hex)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#>   predict        p0         p1
#> 1       0 0.8730189 0.12698107
#> 2       1 0.3718861 0.62811391
#> 3       0 0.9027830 0.09721696
#> 4       0 0.8133735 0.18662654
#> 5       0 0.9424257 0.05757430
#> 6       1 0.1722434 0.82775659
#> 
#> [380 rows x 3 columns] 

h2o.predict_leaf_node_assignment(prostate.gbm, prostate.hex)

#>   T1.C1 T2.C1 T3.C1 T4.C1 T5.C1 T6.C1 T7.C1 T8.C1 T9.C1 T10.C1 T11.C1 T12.C1
#> 1 LLRRL LLRRL LLRLL LLRRL LLLLR LRLLL LLRLL LRLRL LLRLL  LLLLR  LLRLR  LRLLR
#> 2 RLRLL RLRRL  RLRR  RLRR RLRLL  RLRR RLLRR RLRLL RLRLL  RLLLL    RRL  RLLLL
#> 3 LLRLL LLRLR LLRLL LLRLL LLLRL LRLLR LLRLL LLLLL  LLLR  LLLRL  LLRLL  LRLLL
#> 4 RRLLR  RRLR   RRR RLRLR RLRLR RLRLR RLLRL RRLRR RLRLR  RLLRR  RLLRL  RLLRR
#> 5 LLRLR LLRLR  LLRR LLRLR LLLLR LRRRL  LLRR  LLRR   LRR  LLLLR    LRR  LRRRR
#> 6 RLRLL RLRRL  RLRR   RRL   RRL   RRL  RLRL RLRLL   RRL   RLRL    RRL   RLRL
#>   T13.C1 T14.C1 T15.C1 T16.C1 T17.C1 T18.C1 T19.C1 T20.C1 T21.C1 T22.C1 T23.C1
#> 1  LLRRL   LRLL  LLRLR  LRLLR  LLRLL  LLRLL  RLRLL  RLLLL  LLRLL    LRL    RLL
#> 2    RRL  RLLRR  RLRRR  RLLLL  RLRRL  LRRLL  RLRLR  RLLRL  LRLLL  RLLLL  RRRLR
#> 3  LLRLL  LRLRR  LLLLL  LRLLL  LLRLL  LLRLR  RLLLL  RLLLL  LLRLR  RLLLL  RRLLL
#> 4  RLLRR  RLLRL  RLRRL  RLLRR  RLRRR   LLRR    RRR  RLLRR   LLRR  RLLRR  RRLRR
#> 5    LRR   LRRR  LLLRR  LRRRR  LRRLL  LLRLR  RLLLR    RLR  LLRLR  RLLLL  RRLLR
#> 6    RRL    RRL  RLRRR    RRL  RLRRL     RR  RLRLR  RLLLL     RR   RLRR  RRRLR
#>   T24.C1 T25.C1 T26.C1 T27.C1 T28.C1 T29.C1 T30.C1 T31.C1 T32.C1 T33.C1 T34.C1
#> 1  RLRLL  LLRRL  RLRRL   LLRL  RLRLL  RLRRL   LLLL  LLLRL  LLLLL   LLRL  LLLRL
#> 2  RLRLR  LRRLL  RLRRL  RLLLL  RLRLR  RLRRL  RRRLL   RLLR  LLLLL  RRLLR  LLLRL
#> 3   RLLL  LLRRL  RLRRL  RLLLL  RLLLL  RLRRL  RLLRL  LLRRL  LLLLL  RLLRL  LLLLL
#> 4  RRRRR  LLRRR  RRRLR  RLLRR  RLRRR  RRRRR  RLLRR  RRRLR  LLLLL  RRRRR  LLLRL
#> 5  RRRRR  LLRRL  RRRLL  RLLLL  RLLRR  RRRRR  RLLRL  LRRLR     LR  RLRRL     LR
#> 6  RLRLR     RL  RLRRL   RLRR  RLRLR  RLRRL    RRL   RLLL  LLLRR  RRLLR  LLLRL
#>   T35.C1 T36.C1 T37.C1 T38.C1 T39.C1 T40.C1 T41.C1 T42.C1 T43.C1 T44.C1 T45.C1
#> 1  RLLLL  LLLLL  RLRRL   LLRL   LLRL   LLRL  RLLLL  LLLLR  RLLLR  LLLLL  LLLLL
#> 2  RRLLL   LLRL  RLRRL   RLLR  RRRLL  RLLLL  RLLLL  LRRLL  RLRLL  LLLLR  LLLLL
#> 3  RRLLL   LLRL  RLRRL   LRLR  RLLLR  RLLLL  RLLLL  LRLLL  RLRLL  LLLLR  LLLLL
#> 4   RRLR   LRLR  RRRRR  RRRLR  RLRRR  RLLRR  RLLLL   LRLR   RLRR  LLLLR  LLLRL
#> 5   RRRR  LLLRR  RRRRL  LRRRL  RLLRR  RRRRL     RR  LRLLL  RLRLL   LLLR      R
#> 6  RRLLL   LLRL  RLRRL   RLLL    RRL  RLLLL  RLLLL  LRRLL  RLRLL  LLLLR  LLLLL
#>   T46.C1 T47.C1 T48.C1 T49.C1 T50.C1
#> 1    LRL  RLLRR  LLLLL  LLLRR  LLLLR
#> 2  RRLLL  RLLRR  LLLLL    RLL  LLLLR
#> 3  RRLLL  RLLRR  LLLLL    RLL  LLLLR
#> 4   RRLR  RLRRR    RRL    LRL    LRR
#> 5   RRRR     RR  LLLLL  LLRRR  LLLLR
#> 6  RRLLL  RLLRR  LLLLL    RLL     RL
#> 
#> [380 rows x 50 columns]