Data H2OFrame Creation in H2O

Creates a data frame in H2O with real-valued, categorical, integer, and binary columns specified by the user.

h2o.createFrame(rows = 10000, cols = 10, randomize = TRUE, value = 0,
  real_range = 100, categorical_fraction = 0.2, factors = 100,
  integer_fraction = 0.2, integer_range = 100, binary_fraction = 0.1,
  binary_ones_fraction = 0.02, time_fraction = 0, string_fraction = 0,
  missing_fraction = 0.01, response_factors = 2, has_response = FALSE,
  seed, seed_for_column_types)

Arguments

rows
The number of rows of data to generate.

cols
The number of columns of data to generate. Excludes the response column if has_response = TRUE.

randomize
A logical value indicating whether data values should be randomly generated. This must be TRUE if either categorical_fraction or integer_fraction is non-zero.

value
If randomize = FALSE, then all real-valued entries will be set to this value.

real_range
The range of randomly generated real values.

categorical_fraction
The fraction of total columns that are categorical.

factors
The number of (unique) factor levels in each categorical column.

integer_fraction
The fraction of total columns that are integer-valued.

integer_range
The range of randomly generated integer values.

binary_fraction
The fraction of total columns that are binary-valued.

binary_ones_fraction
The fraction of values in a binary column that are set to 1.

time_fraction
The fraction of randomly created date/time columns.

string_fraction
The fraction of randomly created string columns.

missing_fraction
The fraction of total entries in the data frame that are set to NA.

response_factors
If has_response = TRUE, then this is the number of factor levels in the response column.

has_response
A logical value indicating whether an additional response column should be pre-pended to the final H2O data frame. If set to TRUE, the total number of columns will be cols+1.

seed
A seed used to generate random values when randomize = TRUE.

seed_for_column_types
A seed used to generate random column types when randomize = TRUE.

Value

Returns an H2OFrame object.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.98 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

hex <- h2o.createFrame(rows = 1000, cols = 100, categorical_fraction = 0.1,
                       factors = 5, integer_fraction = 0.5, integer_range = 1,
                       has_response = TRUE)

#>   |                                                                              |                                                                      |   0%  |                                                                              |====================                                                  |  29%  |                                                                              |======================================================================| 100%

head(hex)

#>   response C1 C2        C3 C4    C5         C6 C7 C8    C9        C10    C11
#> 1        1  0 -1  87.75281  0 c5.l0 -71.425697  1  1 c9.l0   5.808923 c11.l3
#> 2        1 -1  0 -46.37276  0 c5.l4   1.314431 -1 -1 c9.l4  44.729588 c11.l3
#> 3        0 -1  1 -19.78965  0 c5.l0 -16.926418 -1  1 c9.l3   8.100996 c11.l3
#> 4        0 -1  0 -89.45412  0 c5.l1  50.167984  0  1 c9.l4 -64.946038 c11.l1
#> 5        0  1  0  74.84006  0 c5.l1  56.312024  0  1 c9.l4  60.395594 c11.l4
#> 6        1  1 -1 -48.76102  0 c5.l3 -68.787406  0 -1 c9.l2  48.717113 c11.l3
#>   C12        C13 C14 C15 C16 C17       C18 C19 C20 C21 C22 C23       C24    C25
#> 1   0 -73.109640   1  -1   1   0  97.67699  -1   1   1  -1   1 -21.05910 c25.l4
#> 2   0  42.320251  -1  -1  -1   1 -80.90420   0   0   1   1   0  20.59683 c25.l2
#> 3   0 -55.243930  -1   1   0   0  65.08054   0  -1   0   0  -1 -11.41454   <NA>
#> 4   0  66.205698  -1  -1   1   0 -78.23650   0   0   1   0  -1 -76.62820 c25.l1
#> 5   0   5.412963  -1   1  -1  -1  51.12342   0  -1   0   1   1  91.13922 c25.l2
#> 6   0  69.859488   1   0   1  -1  52.27245  -1  -1   0   0   0 -74.24864 c25.l2
#>   C26 C27        C28 C29 C30       C31 C32        C33       C34    C35 C36
#> 1   1  -1  -4.872610   1   0 -33.58093   0 -92.559438  48.82797 c35.l2   1
#> 2 NaN   0  60.850767   1   0  35.86271   1  -7.191204 -79.52337 c35.l1  -1
#> 3   0   1  78.515767  -1  -1  73.79734   0  66.353209 -93.17586 c35.l4   0
#> 4   0   0  -5.772462  -1   1  70.41048   0   7.225676  10.59825 c35.l2  -1
#> 5   1   1 -59.544774  -1   0 -13.91347   1  75.156291  68.23981 c35.l4   1
#> 6   1   0  27.847567   1  -1 -81.29198  -1 -48.769966 -64.99223 c35.l1   0
#>          C37 C38 C39    C40        C41       C42 C43 C44 C45 C46       C47
#> 1  79.970919   0  -1 c40.l2 -45.917952 -31.08539   0   0   0   0 -56.82841
#> 2  54.994208  -1   1 c40.l2  -3.878275  49.13545   0   0   1   0 -16.59774
#> 3 -82.870919   0   0 c40.l3 -21.905464  35.39073   0   1  -1   0 -94.48263
#> 4  62.536989   1   1 c40.l3  33.847226  52.72721   0   0  -1   0 -31.54193
#> 5  94.538830   0   0 c40.l0 -69.008894 -35.96374   0   0   1   0  55.01813
#> 6  -8.033655  -1  -1 c40.l3 -54.925860  68.24686   0   0 NaN   0  94.72409
#>         C48 C49 C50       C51 C52 C53       C54       C55 C56 C57 C58    C59
#> 1  53.39958   0   0  23.35854   1  -1  50.67585 -31.35095  -1   1  -1 c59.l0
#> 2 -55.73904   0   0  41.57045  -1   1 -12.54783  74.69268   1  -1   0 c59.l3
#> 3 -97.35396   1   0  38.56140  -1   0  38.31048 -13.94393  -1  -1   1 c59.l3
#> 4  19.14822  -1   0  66.69927  -1   1 -22.51774  37.97901   0   0  -1 c59.l2
#> 5  95.95742  -1   0 -70.05963  -1  -1  64.47980  83.91011  -1   0   0 c59.l2
#> 6 -56.54979  -1   0 -94.34010   1   1 -59.43368 -63.41442   1   0  -1 c59.l4
#>   C60    C61 C62       C63 C64 C65         C66       C67        C68 C69 C70
#> 1   0 c61.l4   0 -20.31835   1   0   4.0800817 -55.28053   6.629391   0   1
#> 2   0 c61.l4   0  21.98659  -1   0   8.3919921 -59.14390  69.067201   0   0
#> 3  -1 c61.l4  -1  39.53341  -1  -1   9.3136927  93.73391 -12.619059   0   0
#> 4   0 c61.l1   0 -23.61831   0  -1 -72.1772512 -61.28638  24.052161   0   0
#> 5  -1 c61.l4   0  13.51562  -1   0   0.7818135  34.07219 -87.744484   0   1
#> 6  -1 c61.l3   0 -40.07767  -1   0  94.4757446 -19.45546  38.417181   0  -1
#>         C71    C72 C73 C74 C75 C76 C77 C78 C79    C80 C81       C82       C83
#> 1 -57.09869 c72.l4  -1 NaN   0   0   1   0  -1 c80.l1   0  27.06871 -24.44537
#> 2 -97.22414 c72.l2  -1   0   1  -1  -1   1   0 c80.l0   0  45.27457 -38.30395
#> 3 -37.44132 c72.l0  -1   0   1  -1   1  -1  -1 c80.l1   0 -34.53231  24.79586
#> 4 -65.60720 c72.l4   1   0  -1   1   1   1  -1 c80.l3   0  96.76759  34.52374
#> 5  92.36193 c72.l0   0   0   0  -1   0 NaN   0 c80.l2   0 -84.36377  88.85625
#> 6 -77.17243 c72.l2   0   0   0   1  -1   0   0 c80.l3   0  31.67631 -31.14305
#>   C84 C85          C86        C87 C88       C89 C90 C91 C92       C93       C94
#> 1  -1   1 -17.58126281  55.381265   0  24.09390   1   0   0  79.84190  40.98463
#> 2  -1   1 -85.28220264  10.543531   1 -79.01172   0   0   0  86.74870  94.68755
#> 3  -1   1 -77.98388570   2.584875   0  13.99757   1   0  -1 -72.93180  71.03580
#> 4  -1   0   0.01046848 -29.452055   0 -50.66200  -1   0  -1  56.15935  62.83868
#> 5  -1  -1 -81.22876588  70.899491  -1  82.91395   0 NaN   0 -20.27518 -91.49214
#> 6   1  -1 -53.48123783 -10.388663  -1 -36.51017   0   0   1 -77.64980 -32.67468
#>   C95 C96 C97 C98 C99 C100
#> 1   1   0   1   0  -1    0
#> 2   1  -1   0   0   1    0
#> 3   0  -1  -1  -1   0   -1
#> 4  -1   1   0  -1  -1   -1
#> 5   1   1   1   0   1    0
#> 6   1   1   1  -1   1    1

summary(hex)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  response C1                 C2                 C3                
#>  1 :503   Min.   :-1.00000   Min.   :-1.00000   Min.   : -99.999  
#>  0 :487   1st Qu.:-1.00000   1st Qu.:-1.00000   1st Qu.: -51.849  
#>  NA: 10   Median : 0.00000   Median : 0.00000   Median :  -2.902  
#>           Mean   : 0.02128   Mean   : 0.04444   Mean   :  -1.031  
#>           3rd Qu.: 1.00000   3rd Qu.: 1.00000   3rd Qu.:  48.836  
#>           Max.   : 1.00000   Max.   : 1.00000   Max.   :  99.378  
#>           NA's   :13         NA's   :10         NA's   :6         
#>  C4                C5         C6                 C7                 
#>  Min.   :0.00000   c5.l1:220  Min.   :-99.9701   Min.   :-1.000000  
#>  1st Qu.:0.00000   c5.l0:205  1st Qu.:-53.1611   1st Qu.:-1.000000  
#>  Median :0.00000   c5.l3:195  Median :  2.6399   Median : 0.000000  
#>  Mean   :0.02012   c5.l4:191  Mean   :  0.2598   Mean   :-0.003027  
#>  3rd Qu.:0.00000   c5.l2:179  3rd Qu.: 51.9966   3rd Qu.: 1.000000  
#>  Max.   :1.00000   NA   : 10  Max.   : 99.8547   Max.   : 1.000000  
#>  NA's   :6                    NA's   :8          NA's   :9          
#>  C8                 C9         C10                 C11        
#>  Min.   :-1.00000   c9.l3:208  Min.   :-99.59660   c11.l1:214 
#>  1st Qu.:-1.00000   c9.l2:206  1st Qu.:-46.31719   c11.l4:209 
#>  Median : 0.00000   c9.l4:201  Median :  0.07781   c11.l3:204 
#>  Mean   : 0.01615   c9.l1:189  Mean   :  0.36157   c11.l0:184 
#>  3rd Qu.: 1.00000   c9.l0:186  3rd Qu.: 47.27100   c11.l2:179 
#>  Max.   : 1.00000   NA   : 10  Max.   : 99.95176   NA    : 10 
#>  NA's   :9                     NA's   :10                     
#>  C12               C13               C14                 C15               
#>  Min.   :0.00000   Min.   :-99.764   Min.   :-1.000000   Min.   :-1.00000  
#>  1st Qu.:0.00000   1st Qu.:-48.452   1st Qu.:-1.000000   1st Qu.:-1.00000  
#>  Median :0.00000   Median :  3.456   Median : 0.000000   Median : 0.00000  
#>  Mean   :0.02424   Mean   :  2.915   Mean   :-0.009091   Mean   :-0.03535  
#>  3rd Qu.:0.00000   3rd Qu.: 57.005   3rd Qu.: 1.000000   3rd Qu.: 1.00000  
#>  Max.   :1.00000   Max.   : 99.118   Max.   : 1.000000   Max.   : 1.00000  
#>  NA's   :10        NA's   :12        NA's   :10          NA's   :10        
#>  C16                 C17                C18                C19                 
#>  Min.   :-1.000000   Min.   :-1.00000   Min.   :-99.8991   Min.   :-1.000e+00  
#>  1st Qu.:-1.000000   1st Qu.:-1.00000   1st Qu.:-50.7190   1st Qu.:-1.000e+00  
#>  Median : 0.000000   Median : 0.00000   Median : -0.6401   Median : 0.000e+00  
#>  Mean   :-0.007107   Mean   : 0.01824   Mean   :  0.6441   Mean   : 7.141e-18  
#>  3rd Qu.: 1.000000   3rd Qu.: 1.00000   3rd Qu.: 50.5871   3rd Qu.: 1.000e+00  
#>  Max.   : 1.000000   Max.   : 1.00000   Max.   : 99.8172   Max.   : 1.000e+00  
#>  NA's   :15          NA's   :13         NA's   :10         NA's   :15          
#>  C20                 C21                C22                 C23               
#>  Min.   :-1.000000   Min.   :-1.00000   Min.   :-1.000000   Min.   :-1.00000  
#>  1st Qu.:-1.000000   1st Qu.:-1.00000   1st Qu.:-1.000000   1st Qu.:-1.00000  
#>  Median : 0.000000   Median : 0.00000   Median : 0.000000   Median : 0.00000  
#>  Mean   :-0.007028   Mean   : 0.04129   Mean   : 0.006042   Mean   :-0.01615  
#>  3rd Qu.: 1.000000   3rd Qu.: 1.00000   3rd Qu.: 1.000000   3rd Qu.: 1.00000  
#>  Max.   : 1.000000   Max.   : 1.00000   Max.   : 1.000000   Max.   : 1.00000  
#>  NA's   :4           NA's   :7          NA's   :7           NA's   :9         
#>  C24               C25         C26                C27               
#>  Min.   :-99.154   c25.l1:222  Min.   :-1.00000   Min.   :-1.00000  
#>  1st Qu.:-45.504   c25.l2:200  1st Qu.:-1.00000   1st Qu.:-1.00000  
#>  Median :  2.780   c25.l3:192  Median : 0.00000   Median : 0.00000  
#>  Mean   :  2.691   c25.l0:189  Mean   :-0.00404   Mean   : 0.02727  
#>  3rd Qu.: 50.866   c25.l4:184  3rd Qu.: 1.00000   3rd Qu.: 1.00000  
#>  Max.   : 99.547   NA    : 13  Max.   : 1.00000   Max.   : 1.00000  
#>  NA's   :9                     NA's   :10         NA's   :10        
#>  C28                C29                C30                C31              
#>  Min.   :-99.6557   Min.   :-1.00000   Min.   :-1.00000   Min.   :-99.786  
#>  1st Qu.:-49.8937   1st Qu.:-1.00000   1st Qu.:-1.00000   1st Qu.:-47.462  
#>  Median :  0.5658   Median : 0.00000   Median : 0.00000   Median :  1.871  
#>  Mean   :  1.0168   Mean   : 0.02618   Mean   :-0.01011   Mean   :  0.528  
#>  3rd Qu.: 50.0290   3rd Qu.: 1.00000   3rd Qu.: 1.00000   3rd Qu.: 46.122  
#>  Max.   : 99.5918   Max.   : 1.00000   Max.   : 1.00000   Max.   : 99.542  
#>  NA's   :12         NA's   :7          NA's   :11         NA's   :10       
#>  C32                C33               C34                  C35        
#>  Min.   :-1.00000   Min.   :-99.978   Min.   :-99.908539   c35.l4:224 
#>  1st Qu.:-1.00000   1st Qu.:-53.104   1st Qu.:-50.646430   c35.l0:198 
#>  Median : 0.00000   Median : -3.648   Median :  0.606066   c35.l3:196 
#>  Mean   : 0.01112   Mean   : -1.847   Mean   :  0.006364   c35.l2:189 
#>  3rd Qu.: 1.00000   3rd Qu.: 50.377   3rd Qu.: 51.759043   c35.l1:187 
#>  Max.   : 1.00000   Max.   : 98.641   Max.   : 99.130282   NA    :  6 
#>  NA's   :11         NA's   :15        NA's   :8                       
#>  C36                C37               C38                C39              
#>  Min.   :-1.00000   Min.   :-99.299   Min.   :-1.00000   Min.   :-1.0000  
#>  1st Qu.:-1.00000   1st Qu.:-47.503   1st Qu.:-1.00000   1st Qu.:-1.0000  
#>  Median : 0.00000   Median :  2.599   Median : 0.00000   Median : 0.0000  
#>  Mean   :-0.01518   Mean   :  1.394   Mean   :-0.01514   Mean   : 0.0131  
#>  3rd Qu.: 1.00000   3rd Qu.: 50.511   3rd Qu.: 1.00000   3rd Qu.: 1.0000  
#>  Max.   : 1.00000   Max.   : 99.916   Max.   : 1.00000   Max.   : 1.0000  
#>  NA's   :12         NA's   :8         NA's   :9          NA's   :8        
#>  C40         C41                C42                C43              
#>  c40.l0:210  Min.   :-99.8594   Min.   :-99.7368   Min.   :0.00000  
#>  c40.l3:210  1st Qu.:-51.5027   1st Qu.:-50.2152   1st Qu.:0.00000  
#>  c40.l2:196  Median : -0.5483   Median :  3.4998   Median :0.00000  
#>  c40.l4:189  Mean   : -0.6138   Mean   :  0.7204   Mean   :0.02321  
#>  c40.l1:185  3rd Qu.: 48.1581   3rd Qu.: 49.0278   3rd Qu.:0.00000  
#>  NA    : 10  Max.   : 99.9617   Max.   : 99.9472   Max.   :1.00000  
#>              NA's   :10         NA's   :7          NA's   :9        
#>  C44               C45                 C46               C47              
#>  Min.   :0.00000   Min.   :-1.000000   Min.   :0.00000   Min.   :-99.983  
#>  1st Qu.:0.00000   1st Qu.:-1.000000   1st Qu.:0.00000   1st Qu.:-48.491  
#>  Median :0.00000   Median : 0.000000   Median :0.00000   Median :  2.600  
#>  Mean   :0.02018   Mean   :-0.006054   Mean   :0.01921   Mean   :  1.035  
#>  3rd Qu.:0.00000   3rd Qu.: 1.000000   3rd Qu.:0.00000   3rd Qu.: 49.693  
#>  Max.   :1.00000   Max.   : 1.000000   Max.   :1.00000   Max.   : 99.984  
#>  NA's   :9         NA's   :9           NA's   :11        NA's   :9        
#>  C48                C49                C50               C51              
#>  Min.   :-99.7292   Min.   :-1.00000   Min.   :0.00000   Min.   :-99.888  
#>  1st Qu.:-50.9289   1st Qu.:-1.00000   1st Qu.:0.00000   1st Qu.:-52.054  
#>  Median :  0.6657   Median : 0.00000   Median :0.00000   Median : -1.424  
#>  Mean   :  0.3233   Mean   : 0.03036   Mean   :0.01933   Mean   : -2.387  
#>  3rd Qu.: 50.1646   3rd Qu.: 1.00000   3rd Qu.:0.00000   3rd Qu.: 48.008  
#>  Max.   : 99.8630   Max.   : 1.00000   Max.   :1.00000   Max.   : 99.836  
#>  NA's   :14         NA's   :12         NA's   :17        NA's   :8        
#>  C52               C53                 C54                C55              
#>  Min.   :-1.0000   Min.   :-1.000000   Min.   :-99.9637   Min.   :-99.413  
#>  1st Qu.:-1.0000   1st Qu.:-1.000000   1st Qu.:-56.0206   1st Qu.:-47.772  
#>  Median : 0.0000   Median : 0.000000   Median :  1.1054   Median :  4.468  
#>  Mean   :-0.0141   Mean   : 0.006091   Mean   :  0.3505   Mean   :  1.593  
#>  3rd Qu.: 1.0000   3rd Qu.: 1.000000   3rd Qu.: 53.0381   3rd Qu.: 50.926  
#>  Max.   : 1.0000   Max.   : 1.000000   Max.   : 99.7776   Max.   : 99.976  
#>  NA's   :7         NA's   :15          NA's   :17         NA's   :11       
#>  C56                 C57                C58                 C59        
#>  Min.   :-1.000000   Min.   :-1.00000   Min.   :-1.000000   c59.l2:205 
#>  1st Qu.:-1.000000   1st Qu.:-1.00000   1st Qu.:-1.000000   c59.l4:201 
#>  Median : 0.000000   Median : 0.00000   Median : 0.000000   c59.l0:199 
#>  Mean   : 0.006073   Mean   : 0.01715   Mean   :-0.009045   c59.l3:194 
#>  3rd Qu.: 1.000000   3rd Qu.: 1.00000   3rd Qu.: 1.000000   c59.l1:193 
#>  Max.   : 1.000000   Max.   : 1.00000   Max.   : 1.000000   NA    :  8 
#>  NA's   :12          NA's   :9          NA's   :5                      
#>  C60                C61         C62                C63                
#>  Min.   :-1.00000   c61.l1:219  Min.   :-1.00000   Min.   :-99.85709  
#>  1st Qu.:-1.00000   c61.l4:215  1st Qu.:-1.00000   1st Qu.:-50.26557  
#>  Median : 0.00000   c61.l0:196  Median : 0.00000   Median : -1.67288  
#>  Mean   : 0.01818   c61.l2:190  Mean   : 0.01107   Mean   : -0.01588  
#>  3rd Qu.: 1.00000   c61.l3:167  3rd Qu.: 1.00000   3rd Qu.: 52.26351  
#>  Max.   : 1.00000   NA    : 13  Max.   : 1.00000   Max.   : 99.90732  
#>  NA's   :10                     NA's   :6          NA's   :4          
#>  C64                 C65                  C66                C67              
#>  Min.   :-1.000000   Min.   :-1.000e+00   Min.   :-99.6488   Min.   :-99.894  
#>  1st Qu.:-1.000000   1st Qu.:-1.000e+00   1st Qu.:-50.8835   1st Qu.:-53.128  
#>  Median : 0.000000   Median : 0.000e+00   Median :  2.0617   Median : -4.964  
#>  Mean   : 0.002033   Mean   :-2.318e-17   Mean   :  0.5543   Mean   : -1.607  
#>  3rd Qu.: 1.000000   3rd Qu.: 1.000e+00   3rd Qu.: 51.4241   3rd Qu.: 51.995  
#>  Max.   : 1.000000   Max.   : 1.000e+00   Max.   : 99.3932   Max.   : 99.960  
#>  NA's   :16          NA's   :12           NA's   :13         NA's   :11       
#>  C68                C69               C70                C71              
#>  Min.   :-99.7320   Min.   :0.00000   Min.   :-1.00000   Min.   :-99.965  
#>  1st Qu.:-51.7653   1st Qu.:0.00000   1st Qu.:-1.00000   1st Qu.:-48.302  
#>  Median :  0.5939   Median :0.00000   Median : 0.00000   Median : -4.027  
#>  Mean   : -0.8035   Mean   :0.02113   Mean   : 0.01913   Mean   : -1.809  
#>  3rd Qu.: 48.7603   3rd Qu.:0.00000   3rd Qu.: 1.00000   3rd Qu.: 45.788  
#>  Max.   : 99.9215   Max.   :1.00000   Max.   : 1.00000   Max.   : 99.697  
#>  NA's   :6          NA's   :6         NA's   :7          NA's   :12       
#>  C72         C73                C74               C75                
#>  c72.l0:225  Min.   :-1.00000   Min.   :0.00000   Min.   :-1.000000  
#>  c72.l4:200  1st Qu.:-1.00000   1st Qu.:0.00000   1st Qu.:-1.000000  
#>  c72.l2:195  Median : 0.00000   Median :0.00000   Median : 0.000000  
#>  c72.l3:191  Mean   : 0.02626   Mean   :0.01714   Mean   : 0.009082  
#>  c72.l1:183  3rd Qu.: 1.00000   3rd Qu.:0.00000   3rd Qu.: 1.000000  
#>  NA    :  6  Max.   : 1.00000   Max.   :1.00000   Max.   : 1.000000  
#>              NA's   :10         NA's   :8         NA's   :9          
#>  C76                C77               C78                C79              
#>  Min.   :-1.00000   Min.   :-1.0000   Min.   :-1.00000   Min.   :-1.0000  
#>  1st Qu.:-1.00000   1st Qu.:-1.0000   1st Qu.:-1.00000   1st Qu.:-1.0000  
#>  Median : 0.00000   Median : 0.0000   Median : 0.00000   Median : 0.0000  
#>  Mean   : 0.03128   Mean   :-0.0202   Mean   : 0.01813   Mean   : 0.0364  
#>  3rd Qu.: 1.00000   3rd Qu.: 1.0000   3rd Qu.: 1.00000   3rd Qu.: 1.0000  
#>  Max.   : 1.00000   Max.   : 1.0000   Max.   : 1.00000   Max.   : 1.0000  
#>  NA's   :9          NA's   :10        NA's   :7          NA's   :11       
#>  C80         C81               C82                C83               
#>  c80.l0:221  Min.   :0.00000   Min.   :-99.7872   Min.   :-99.8442  
#>  c80.l3:203  1st Qu.:0.00000   1st Qu.:-50.2696   1st Qu.:-49.7086  
#>  c80.l1:202  Median :0.00000   Median : -1.5506   Median : -2.9686  
#>  c80.l2:192  Mean   :0.02014   Mean   : -0.1039   Mean   :  0.1865  
#>  c80.l4:175  3rd Qu.:0.00000   3rd Qu.: 50.5627   3rd Qu.: 49.8635  
#>  NA    :  7  Max.   :1.00000   Max.   : 99.8806   Max.   : 99.8993  
#>              NA's   :7         NA's   :6          NA's   :13        
#>  C84                C85                C86                 C87              
#>  Min.   :-1.00000   Min.   :-1.00000   Min.   :-99.77936   Min.   :-99.855  
#>  1st Qu.:-1.00000   1st Qu.:-1.00000   1st Qu.:-53.65201   1st Qu.:-48.869  
#>  Median : 0.00000   Median : 0.00000   Median : -1.33441   Median :  4.113  
#>  Mean   :-0.05131   Mean   : 0.04053   Mean   : -0.03536   Mean   :  1.584  
#>  3rd Qu.: 1.00000   3rd Qu.: 1.00000   3rd Qu.: 52.58067   3rd Qu.: 51.108  
#>  Max.   : 1.00000   Max.   : 1.00000   Max.   : 99.90613   Max.   : 99.700  
#>  NA's   :6          NA's   :13         NA's   :5           NA's   :9        
#>  C88                C89                 C90               C91              
#>  Min.   :-1.00000   Min.   :-99.91107   Min.   :-1.0000   Min.   :0.00000  
#>  1st Qu.:-1.00000   1st Qu.:-51.41437   1st Qu.:-1.0000   1st Qu.:0.00000  
#>  Median : 0.00000   Median : -0.81998   Median : 0.0000   Median :0.00000  
#>  Mean   : 0.03033   Mean   :  0.06611   Mean   :-0.0191   Mean   :0.02427  
#>  3rd Qu.: 1.00000   3rd Qu.: 50.57354   3rd Qu.: 1.0000   3rd Qu.:0.00000  
#>  Max.   : 1.00000   Max.   : 99.86936   Max.   : 1.0000   Max.   :1.00000  
#>  NA's   :11         NA's   :4           NA's   :5         NA's   :11       
#>  C92                C93                C94                C95               
#>  Min.   :-1.00000   Min.   :-99.9676   Min.   :-99.8194   Min.   :-1.00000  
#>  1st Qu.:-1.00000   1st Qu.:-54.1200   1st Qu.:-46.8673   1st Qu.:-1.00000  
#>  Median : 0.00000   Median :  0.5176   Median :  0.2900   Median : 0.00000  
#>  Mean   : 0.01518   Mean   : -0.4656   Mean   :  0.8095   Mean   :-0.04555  
#>  3rd Qu.: 1.00000   3rd Qu.: 52.8578   3rd Qu.: 50.8443   3rd Qu.: 1.00000  
#>  Max.   : 1.00000   Max.   : 99.8042   Max.   : 99.9998   Max.   : 1.00000  
#>  NA's   :12         NA's   :5          NA's   :11         NA's   :12        
#>  C96                 C97                C98               C99               
#>  Min.   :-1.000000   Min.   :-1.00000   Min.   :-1.0000   Min.   :-1.00000  
#>  1st Qu.:-1.000000   1st Qu.:-1.00000   1st Qu.:-1.0000   1st Qu.:-1.00000  
#>  Median : 0.000000   Median : 0.00000   Median : 0.0000   Median : 0.00000  
#>  Mean   :-0.005045   Mean   : 0.01316   Mean   : 0.0091   Mean   : 0.04632  
#>  3rd Qu.: 1.000000   3rd Qu.: 1.00000   3rd Qu.: 1.0000   3rd Qu.: 1.00000  
#>  Max.   : 1.000000   Max.   : 1.00000   Max.   : 1.0000   Max.   : 1.00000  
#>  NA's   :9           NA's   :12         NA's   :11        NA's   :7         
#>  C100              
#>  Min.   :-1.00000  
#>  1st Qu.:-1.00000  
#>  Median : 0.00000  
#>  Mean   : 0.02328  
#>  3rd Qu.: 1.00000  
#>  Max.   : 1.00000  
#>  NA's   :12        

hex2 <- h2o.createFrame(rows = 100, cols = 10, randomize = FALSE, value = 5,
                        categorical_fraction = 0, integer_fraction = 0)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

summary(hex2)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  C1            C2            C3            C4            C5           
#>  Min.   :  5   Min.   :  5   Min.   :  5   Min.   :  5   Min.   :  5  
#>  1st Qu.:NaN   1st Qu.:NaN   1st Qu.:NaN   1st Qu.:NaN   1st Qu.:NaN  
#>  Median :NaN   Median :NaN   Median :NaN   Median :NaN   Median :NaN  
#>  Mean   :  5   Mean   :  5   Mean   :  5   Mean   :  5   Mean   :  5  
#>  3rd Qu.:NaN   3rd Qu.:NaN   3rd Qu.:NaN   3rd Qu.:NaN   3rd Qu.:NaN  
#>  Max.   :  5   Max.   :  5   Max.   :  5   Max.   :  5   Max.   :  5  
#>  NA's   :3     NA's   :1     NA's   :1     NA's   :2     NA's   :1    
#>  C6            C7            C8            C9            C10          
#>  Min.   :  5   Min.   :  5   Min.   :  5   Min.   :  5   Min.   :  5  
#>  1st Qu.:NaN   1st Qu.:NaN   1st Qu.:NaN   1st Qu.:NaN   1st Qu.:NaN  
#>  Median :NaN   Median :NaN   Median :NaN   Median :NaN   Median :NaN  
#>  Mean   :  5   Mean   :  5   Mean   :  5   Mean   :  5   Mean   :  5  
#>  3rd Qu.:NaN   3rd Qu.:NaN   3rd Qu.:NaN   3rd Qu.:NaN   3rd Qu.:NaN  
#>  Max.   :  5   Max.   :  5   Max.   :  5   Max.   :  5   Max.   :  5  
#>  NA's   :1     NA's   :2     NA's   :1     NA's   :1     NA's   :1