Feature Generation via H2O Deep Learning or DeepWater Model

Extract the non-linear feature from an H2O data set using an H2O deep learning model.

h2o.deepfeatures(object, data, layer)

Arguments

object
An H2OModel object that represents the deep learning model to be used for feature extraction.

data
An H2OFrame object.

layer
Index (for DeepLearning, integer) or Name (for DeepWater, String) of the hidden layer to extract

Value

Returns an H2OFrame object with as many features as the number of units in the hidden layer of the specified index.

See also

link{h2o.deeplearning} for making H2O Deep Learning models. link{h2o.deepwater} for making H2O DeepWater models.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.84 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath = system.file("extdata", "prostate.csv", package = "h2o")
prostate.hex = h2o.importFile(path = prosPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

prostate.dl = h2o.deeplearning(x = 3:9, y = 2, training_frame = prostate.hex,
                               hidden = c(100, 200), epochs = 5)

#>   |                                                                              |                                                                      |   0%  |                                                                              |============================                                          |  40%  |                                                                              |======================================================================| 100%

prostate.deepfeatures_layer1 = h2o.deepfeatures(prostate.dl, prostate.hex, layer = 1)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

prostate.deepfeatures_layer2 = h2o.deepfeatures(prostate.dl, prostate.hex, layer = 2)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

head(prostate.deepfeatures_layer1)

#>     DF.L1.C1  DF.L1.C2  DF.L1.C3  DF.L1.C4   DF.L1.C5   DF.L1.C6  DF.L1.C7
#> 1 0.26194861 0.5106790 0.4411528 0.3412370 0.35652669 0.61477098 0.2347755
#> 2 0.43080466 1.3710128 0.8951324 0.5657259 1.20718645 0.15028624 0.6687946
#> 3 0.41272998 1.2487790 0.7809292 1.0626187 0.53185744 0.00000000 0.6789289
#> 4 0.03827464 0.8347868 0.2292304 0.3353654 0.02240105 0.00000000 1.5660580
#> 5 0.86023044 0.0000000 0.2739298 1.0629305 0.07966208 0.66510322 0.7828144
#> 6 0.60807782 1.5258135 0.7868311 0.4914526 1.44800045 0.03055112 0.7207875
#>    DF.L1.C8  DF.L1.C9  DF.L1.C10 DF.L1.C11 DF.L1.C12 DF.L1.C13  DF.L1.C14
#> 1 0.7898566 0.5026646 0.84319392 0.1774688 0.5129879 0.7361809 0.22233948
#> 2 0.5703027 0.9612336 0.38608583 1.2499061 0.2348775 0.5098597 1.12353327
#> 3 0.4850872 0.6447695 1.06231961 0.9786833 0.2273578 0.2672496 0.71373272
#> 4 0.0000000 0.3673409 0.03486342 1.3756996 0.4453548 0.4380750 0.05156255
#> 5 0.0000000 0.1422243 0.50525193 0.1771863 0.7263337 0.0000000 0.61608931
#> 6 0.4015894 0.8819348 0.18787174 1.3497608 0.1111497 0.5903850 1.17609605
#>   DF.L1.C15  DF.L1.C16  DF.L1.C17 DF.L1.C18 DF.L1.C19 DF.L1.C20 DF.L1.C21
#> 1 0.5469347 0.38140635 0.54180905 0.1063642 0.6320868 0.4500296 0.6896160
#> 2 0.7402958 0.01892915 0.81799647 0.8476386 0.0000000 0.5621489 0.9236719
#> 3 0.8562448 0.00000000 0.79787001 0.8060660 0.0000000 0.3154780 0.6024010
#> 4 0.2752798 0.96923834 0.00000000 0.9791675 0.0000000 0.0000000 0.0000000
#> 5 0.4069045 0.54514722 0.08750631 1.0110116 0.4390364 0.5545027 0.0000000
#> 6 0.7197298 0.04502870 0.94378816 0.9387644 0.0000000 0.7904509 0.9462635
#>   DF.L1.C22 DF.L1.C23  DF.L1.C24  DF.L1.C25 DF.L1.C26  DF.L1.C27 DF.L1.C28
#> 1 0.4622459 0.5682982 0.70239942 0.71201188 0.3266030 0.20385261 0.5308741
#> 2 0.8168494 0.3426150 0.00000000 0.05680309 1.0721402 0.15992779 1.1472226
#> 3 1.0535498 0.2889702 0.00000000 0.17648302 0.5065675 0.21912647 1.2781347
#> 4 0.3933526 0.1983936 0.39110157 0.97275437 0.5410383 0.57822367 0.2953835
#> 5 0.6787172 0.1887831 0.06095204 0.36531658 0.0000000 0.79890015 0.2695031
#> 6 0.9884801 0.5658658 0.00000000 0.10117775 1.2933463 0.02536982 1.1291392
#>   DF.L1.C29 DF.L1.C30 DF.L1.C31 DF.L1.C32 DF.L1.C33 DF.L1.C34 DF.L1.C35
#> 1 0.6132215 0.5116440 0.2843326 0.3353043 0.3947234 0.3427702 0.1253710
#> 2 0.0000000 0.7155079 0.3621937 0.2258558 0.4378109 0.4033111 0.3090162
#> 3 0.1206704 0.8731354 0.5380887 0.0000000 0.0000000 0.2641892 0.1042776
#> 4 0.0000000 0.3230758 0.4034452 1.0662940 0.0000000 0.0000000 0.7210381
#> 5 0.6746432 0.7252362 0.8204490 0.2416422 0.2453218 0.2191709 0.5291220
#> 6 0.0000000 0.5498661 0.2829551 0.3881669 0.5031514 0.5206710 0.5049717
#>     DF.L1.C36  DF.L1.C37 DF.L1.C38  DF.L1.C39 DF.L1.C40 DF.L1.C41  DF.L1.C42
#> 1 0.699648802 0.52759339 0.2630695 0.78032903 0.5077189 0.1439690 0.88074333
#> 2 0.000000000 0.41925329 0.1955342 0.46448033 0.3803183 0.7657101 0.08754385
#> 3 0.296526523 0.05311684 0.0000000 0.04313985 0.6433066 0.6055545 0.66352931
#> 4 0.004550451 0.80989217 0.8990867 0.09151448 0.1086393 0.6899083 0.00000000
#> 5 0.224366126 0.26851351 0.4189918 0.00000000 0.8205816 0.5327894 0.90517140
#> 6 0.049710654 0.34813299 0.1690166 0.58046556 0.4360850 0.8443798 0.00000000
#>   DF.L1.C43 DF.L1.C44 DF.L1.C45 DF.L1.C46 DF.L1.C47 DF.L1.C48 DF.L1.C49
#> 1 0.5453987 0.3672468 0.3027342 0.6961278 0.6778881 0.3473646 0.7272235
#> 2 0.9552625 1.2779357 0.2049392 1.2530363 0.6689778 0.8650190 0.1464998
#> 3 0.9958180 0.8942793 0.2577085 1.4948870 1.1328070 0.7270948 0.6039107
#> 4 0.0000000 0.1877410 0.8226683 0.0000000 0.0000000 0.8130467 0.7400544
#> 5 0.4488273 0.5633612 0.7632955 0.4238315 0.5031675 0.0737292 0.5700643
#> 6 0.7982311 1.4425427 0.1329135 1.0397571 0.7070393 0.8139719 0.1189713
#>    DF.L1.C50 DF.L1.C51 DF.L1.C52 DF.L1.C53 DF.L1.C54 DF.L1.C55 DF.L1.C56
#> 1 0.53539666 0.3431884 0.4319967 0.3846209 0.7787118 0.2595773 0.4539771
#> 2 0.00000000 0.9365582 0.7966211 0.5565560 0.1704920 0.5735374 0.4847626
#> 3 0.03127991 0.4306483 0.7832443 0.3212837 0.2852597 0.7791664 0.5596340
#> 4 0.51726136 1.1536573 1.1631788 1.2220477 0.0000000 1.0123620 1.0222367
#> 5 0.96202316 0.0000000 0.1684526 0.1296455 0.2816554 0.7061539 0.6064674
#> 6 0.00000000 1.0352415 0.8462437 0.7124361 0.2141111 0.5573137 0.3807283
#>   DF.L1.C57 DF.L1.C58 DF.L1.C59 DF.L1.C60 DF.L1.C61  DF.L1.C62 DF.L1.C63
#> 1 0.1942541 0.2964332 0.5529306 0.7415254 0.5611515 0.58401788 0.2349069
#> 2 0.8920150 1.0552589 0.2231178 0.0425638 0.0000000 0.17297286 1.0768570
#> 3 0.6546783 1.0561067 0.3183232 0.6325822 0.6091251 0.00000000 0.7937918
#> 4 0.3885393 0.3777161 0.8558893 0.4705274 0.2861438 0.81819588 0.3478811
#> 5 0.4180454 0.4938485 0.2888772 0.7299741 1.0678277 0.50340260 0.4542338
#> 6 0.9325873 1.1596755 0.2756820 0.0000000 0.0000000 0.07455312 1.1970625
#>    DF.L1.C64 DF.L1.C65 DF.L1.C66 DF.L1.C67  DF.L1.C68 DF.L1.C69 DF.L1.C70
#> 1 0.61151932 0.4232045 0.7249453 0.5407987 0.75101039 0.2374673 0.4459582
#> 2 0.56284911 0.4022704 0.8232093 0.0000000 0.56180948 0.0000000 0.7230671
#> 3 0.75274657 0.6225159 1.2564801 0.0000000 0.98494006 0.0000000 0.6644961
#> 4 0.00000000 0.0000000 0.1288125 0.3859961 0.07975793 1.0255342 0.0000000
#> 5 0.03039595 0.8295162 0.3269041 0.7408503 0.34436250 0.7953810 0.4885908
#> 6 0.48148217 0.5262610 0.8558263 0.0000000 0.55440313 0.0000000 0.9128595
#>   DF.L1.C71 DF.L1.C72  DF.L1.C73 DF.L1.C74  DF.L1.C75 DF.L1.C76 DF.L1.C77
#> 1 0.2141703 0.1325993 0.47554857 0.3542797 0.34610894 0.2068090 0.3390523
#> 2 0.6627735 1.0882570 0.47065653 0.3108553 0.15130079 1.1693010 0.2285754
#> 3 0.3981855 0.7311670 0.28896676 0.4005142 0.20297131 0.4886387 0.1062270
#> 4 1.5504674 1.1803812 0.50483947 0.2886621 0.33911166 0.4034762 0.5233751
#> 5 0.2763146 0.5956696 0.02909304 0.4719718 0.97425978 0.1223568 0.3666532
#> 6 0.4195845 0.9628704 0.66097282 0.2984090 0.03463958 1.3487720 0.4483844
#>    DF.L1.C78  DF.L1.C79 DF.L1.C80 DF.L1.C81 DF.L1.C82 DF.L1.C83 DF.L1.C84
#> 1 0.26750706 0.17775240 0.5557805 0.5046493 0.5543707 0.4235369 0.3007827
#> 2 0.00000000 0.00000000 0.8193639 0.9179962 0.0000000 0.2577737 0.1267554
#> 3 0.07519914 0.00000000 1.0763865 0.8149939 0.0000000 0.2266625 0.0000000
#> 4 0.98847163 0.86230484 0.0000000 0.5298182 0.0000000 0.8968754 0.6750704
#> 5 1.03534553 0.71310934 0.5129626 0.2147965 0.8248572 0.2428781 0.1060456
#> 6 0.00000000 0.05123353 0.6472175 0.8457055 0.0000000 0.4169599 0.3868504
#>   DF.L1.C85 DF.L1.C86 DF.L1.C87 DF.L1.C88 DF.L1.C89 DF.L1.C90 DF.L1.C91
#> 1 0.3724126 0.5954447 0.6552705 0.7553117 0.4232865 0.5066642 0.2805735
#> 2 0.6735379 0.8741186 0.2773632 0.7295850 0.9979936 0.9271522 0.8611294
#> 3 0.7379749 1.0446965 0.4403824 1.4215457 1.2676014 1.0404993 0.2301246
#> 4 1.3700281 0.1579266 0.0000000 0.0000000 0.7294668 0.4712848 0.5175163
#> 5 0.5434631 0.3978358 0.5538001 0.6649066 0.5342335 0.1884847 0.0000000
#> 6 0.6631623 0.9111402 0.2558397 0.6539776 1.1474422 0.8759005 1.0403529
#>   DF.L1.C92 DF.L1.C93 DF.L1.C94 DF.L1.C95 DF.L1.C96 DF.L1.C97 DF.L1.C98
#> 1 0.4463365 0.3020728 0.5942079 0.3293892 0.4103112 0.5258869 0.4993814
#> 2 0.3889736 0.0000000 0.4680219 1.2380426 0.2374613 0.3272063 0.8752812
#> 3 0.6024695 0.0000000 0.3889754 1.1002525 0.1609394 0.4898885 0.6650721
#> 4 0.6669002 0.9842037 0.0000000 0.9894664 1.0422754 0.6082622 0.1863148
#> 5 0.6925643 0.7797452 0.4134861 0.2533909 0.1485230 0.2683317 0.6302923
#> 6 0.2252888 0.0000000 0.6658298 1.3598968 0.3093559 0.2160083 1.0352498
#>   DF.L1.C99 DF.L1.C100
#> 1 0.6192207  0.4615438
#> 2 0.4410434  0.7409357
#> 3 0.7138210  0.4135490
#> 4 0.0000000  0.4365762
#> 5 0.7290866  0.0142248
#> 6 0.3113099  0.8774521

head(prostate.deepfeatures_layer2)

#>    DF.L2.C1  DF.L2.C2 DF.L2.C3 DF.L2.C4  DF.L2.C5  DF.L2.C6  DF.L2.C7  DF.L2.C8
#> 1 0.6538416 0.8657092        0 1.117056 0.7376546 0.3562750 1.0642917 0.7104452
#> 2 1.1915890 0.7610736        0 1.312743 1.0085719 0.3026494 0.9723069 0.4090570
#> 3 1.2547336 0.8619579        0 1.302335 0.8867038 0.1855395 1.1536559 0.6806632
#> 4 0.9994148 0.5113625        0 1.008016 1.4969494 1.2522968 0.4850179 0.3358678
#> 5 1.0280237 1.0619076        0 1.110988 0.8112293 0.3571773 0.9807775 0.0000000
#>    DF.L2.C9 DF.L2.C10 DF.L2.C11 DF.L2.C12 DF.L2.C13 DF.L2.C14 DF.L2.C15
#> 1 1.2228637 0.4825533 0.0000000 1.2816111 0.6996583 0.2190777 0.4765033
#> 2 0.9843043 0.7494075 0.0000000 0.3840901 0.5494587 0.1123328 0.8103155
#> 3 0.7236864 0.4488775 0.0000000 0.9746548 0.4536891 0.0000000 0.8740985
#> 4 1.1598477 0.8660611 0.1859283 0.6164953 0.7786332 0.0000000 1.7240814
#> 5 0.8483402 0.3984079 0.0000000 1.3075125 0.4886257 0.0000000 1.0831920
#>   DF.L2.C16  DF.L2.C17 DF.L2.C18 DF.L2.C19 DF.L2.C20 DF.L2.C21 DF.L2.C22
#> 1 0.7780064 0.69416169 0.6341612 1.5147792 1.0075141 1.2435352 1.1655943
#> 2 0.7375879 0.00000000 0.1570304 1.0587372 0.4523208 0.8347635 0.8109605
#> 3 0.5571926 0.00000000 0.5459973 1.2391095 0.6006494 0.9518656 0.8566560
#> 4 0.9497558 0.00000000 0.2901777 1.6299783 0.6515454 0.8863054 1.4281964
#> 5 0.5389586 0.04752197 0.5645077 2.1147019 1.0704150 1.2580064 1.3299995
#>   DF.L2.C23 DF.L2.C24 DF.L2.C25  DF.L2.C26 DF.L2.C27 DF.L2.C28 DF.L2.C29
#> 1 0.3797811 1.1493696 0.7543977 1.03081067 0.8496338 0.4232502 1.2556873
#> 2 0.0000000 0.0000000 0.9781476 0.32539990 0.6149834 0.4469463 1.3896711
#> 3 0.0000000 0.2830837 1.2705230 0.99247734 0.2016531 0.3147510 1.4464430
#> 4 0.0000000 0.3746954 0.1077996 0.43355575 0.3859872 0.5141272 1.1019425
#> 5 0.1183534 0.4701192 0.9952654 0.99563842 0.3827601 0.4486889 0.9032241
#>   DF.L2.C30  DF.L2.C31 DF.L2.C32 DF.L2.C33 DF.L2.C34 DF.L2.C35 DF.L2.C36
#> 1 0.9405638 0.00000000  1.192077 0.6171322 1.1072485 0.4950882         0
#> 2 0.6688664 0.37155415  1.610844 0.6094385 1.0974636 0.1911893         0
#> 3 0.9543241 0.05718986  1.549619 0.3784238 0.8753253 0.4191145         0
#> 4 1.1936705 0.21276042  1.277518 0.3922528 0.5836025 0.6482422         0
#> 5 1.1724951 0.00000000  1.015229 0.5688311 1.1378109 0.7535679         0
#>   DF.L2.C37 DF.L2.C38 DF.L2.C39 DF.L2.C40 DF.L2.C41 DF.L2.C42  DF.L2.C43
#> 1         0 0.9774738 1.2366161 0.0000000 0.8305965 0.6002921 0.46476121
#> 2         0 0.8241056 0.8138960 0.4846680 1.2456130 0.7100249 0.10825300
#> 3         0 0.8274868 0.8966806 0.6243996 1.3397465 0.6068259 0.57563487
#> 4         0 0.8880508 1.0185620 0.0000000 0.2301627 0.3596817 0.00000000
#> 5         0 1.0136200 1.3443889 0.4145148 0.5001962 0.4631075 0.37648241
#>   DF.L2.C44 DF.L2.C45 DF.L2.C46 DF.L2.C47 DF.L2.C48 DF.L2.C49  DF.L2.C50
#> 1 0.0296598 0.9084035 0.0000000  1.352338 0.7942950 0.8944268 0.05629283
#> 2 0.0000000 0.4565931 0.0000000  1.584657 0.7726121 0.5947260 0.30061204
#> 3 0.0000000 0.6569921 0.0000000  1.554484 1.0330740 0.8850536 0.17691163
#> 4 0.3146877 0.3532507 0.4201740  1.381748 0.7033758 0.6668944 0.00000000
#> 5 0.1321377 0.8403262 0.1382156  1.215009 1.0625696 0.9596074 0.00000000
#>    DF.L2.C51 DF.L2.C52  DF.L2.C53  DF.L2.C54 DF.L2.C55 DF.L2.C56  DF.L2.C57
#> 1 0.00000000         0 0.21419287 0.01770775 1.5217926 0.1398141 0.88439292
#> 2 0.00000000         0 0.09632396 0.00000000 0.8222542 0.7161949 0.03730048
#> 3 0.01912219         0 0.35558914 0.07287231 0.8776634 0.2720573 0.60142794
#> 4 0.00000000         0 1.18261174 0.52446848 1.6349631 0.9221575 0.00000000
#> 5 0.00000000         0 1.01604386 0.00000000 1.2729408 0.6436060 0.73680469
#>   DF.L2.C58 DF.L2.C59  DF.L2.C60 DF.L2.C61 DF.L2.C62 DF.L2.C63 DF.L2.C64
#> 1         0 0.7366515 0.54887234 0.4740660 1.1210276 1.5148575 0.9703314
#> 2         0 0.9449041 0.00000000 0.2635289 2.1470669 1.2270221 0.9428789
#> 3         0 0.9431161 0.08638416 0.4952709 1.7036467 1.4574857 1.2044351
#> 4         0 0.9515233 0.80134648 0.0000000 0.9138093 0.9188766 0.6451632
#> 5         0 0.7014905 0.11722480 0.2949436 0.2628560 1.2829501 0.8203020
#>   DF.L2.C65 DF.L2.C66 DF.L2.C67  DF.L2.C68 DF.L2.C69 DF.L2.C70 DF.L2.C71
#> 1 0.5185607 0.7588880  1.472944 0.66443735 0.7261749 0.5953232 1.2037251
#> 2 1.2802656 0.6801125  1.885972 0.04181098 0.5485773 0.4860200 1.0816377
#> 3 1.2249976 0.9318894  1.651478 0.38604724 0.2681971 0.6605725 1.1972109
#> 4 0.8031147 1.2197367  1.124983 0.00000000 0.9819568 0.8774681 1.0927698
#> 5 0.9125084 0.8228217  1.067479 0.51351446 0.8802325 0.5063981 0.7775973
#>    DF.L2.C72 DF.L2.C73 DF.L2.C74 DF.L2.C75 DF.L2.C76   DF.L2.C77 DF.L2.C78
#> 1 0.44900931 0.7247271 0.8946261 1.1034649 0.7736099 0.000000000 0.5899818
#> 2 0.27825753 1.1560785 0.9781402 0.7362815 1.5790377 0.007495526 0.2300682
#> 3 0.49181001 0.9219584 0.9820093 0.8900212 1.1476111 0.000000000 0.2931641
#> 4 0.32093842 1.1003609 1.1850895 1.1836516 0.8061900 0.340347042 1.2715743
#> 5 0.44990468 0.5562074 1.3203754 1.0890173 0.3128226 0.000000000 0.9972249
#>   DF.L2.C79 DF.L2.C80 DF.L2.C81 DF.L2.C82 DF.L2.C83 DF.L2.C84  DF.L2.C85
#> 1 0.6445871 1.3089654 0.6262664  1.308582         0  1.223305 0.71427999
#> 2 0.0000000 0.8160545 1.0733955  1.112302         0  1.275969 0.36788769
#> 3 0.1989852 0.7475443 0.9729626  1.549010         0  1.576129 0.67173995
#> 4 0.0000000 0.8026321 0.1847145  1.649042         0  1.785223 0.01687173
#> 5 0.4311609 0.8572082 0.3425309  1.317398         0  1.646718 0.17167871
#>    DF.L2.C86 DF.L2.C87 DF.L2.C88  DF.L2.C89 DF.L2.C90 DF.L2.C91 DF.L2.C92
#> 1 0.00000000 0.9344131 1.0656441 0.13123376 0.9620080 0.2070940 0.6911139
#> 2 0.00000000 0.4844184 0.9424875 0.00000000 0.8934321 0.0000000 0.7327565
#> 3 0.00000000 0.6597816 0.9711736 0.00000000 0.7692100 0.0000000 0.7482321
#> 4 0.06238159 0.5034240 0.1512614 0.17416493 0.7816657 0.2139638 0.7171681
#> 5 0.00000000 0.7884990 0.5502392 0.06975427 1.0368137 0.2407858 0.8812431
#>   DF.L2.C93 DF.L2.C94 DF.L2.C95 DF.L2.C96 DF.L2.C97 DF.L2.C98 DF.L2.C99
#> 1 0.9749311  1.160696  0.000000 0.0000000 1.3712601 0.5499773  1.425448
#> 2 1.1007768  1.586044  0.000000 0.3719197 1.0538044 0.6445943  1.113298
#> 3 0.5531689  1.682847  0.000000 0.4297744 1.4234324 0.5628262  1.177404
#> 4 0.9496282  1.141526  0.000000 0.0000000 0.9550672 0.1288212  1.321883
#> 5 0.6288972  1.270537  0.324129 0.2843502 1.3134501 0.4100083  1.358532
#>   DF.L2.C100 DF.L2.C101 DF.L2.C102 DF.L2.C103 DF.L2.C104 DF.L2.C105 DF.L2.C106
#> 1  0.6072094  0.7923481  0.0000000   1.684514  1.0937397  0.3699206 0.74468574
#> 2  0.6158525  0.4610919  0.0000000   2.101250  0.8152430  0.7638433 0.05250258
#> 3  0.5459863  0.9954478  0.0000000   1.701176  0.5877417  0.7395534 0.09057927
#> 4  0.6842335  1.1490987  0.1523032   1.598044  1.2111832  0.2447003 0.34779966
#> 5  0.5003226  1.1792698  0.0000000   1.719046  1.0435764  0.3413306 0.29095791
#>   DF.L2.C107 DF.L2.C108 DF.L2.C109 DF.L2.C110 DF.L2.C111 DF.L2.C112 DF.L2.C113
#> 1 0.51625158  0.8560127  0.9862538  0.0000000          0   1.481948  1.1105953
#> 2 0.08549379  1.5611946  1.1075537  0.0000000          0   2.201934  0.8793948
#> 3 0.04524581  1.3306066  1.3656358  0.0000000          0   2.004000  0.7885367
#> 4 0.14934168  0.8436412  1.3246823  0.3079333          0   1.330970  1.5272395
#> 5 0.51970914  1.0667976  1.1943679  0.0000000          0   1.216805  0.9863802
#>   DF.L2.C114 DF.L2.C115 DF.L2.C116 DF.L2.C117 DF.L2.C118 DF.L2.C119 DF.L2.C120
#> 1 0.38158715  0.0000000  0.8048023  0.1854080  0.8849421   1.127602   1.468792
#> 2 0.00000000  0.4587677  0.7704076  0.1015204  0.9410624   1.827032   1.724540
#> 3 0.00000000  0.5383148  1.0034480  0.2634821  1.1494982   1.841985   1.705400
#> 4 0.34557441  0.5140435  0.6923128  0.4866748  1.1096787   1.480583   1.080272
#> 5 0.03325183  0.5611853  1.0078391  0.5210694  0.9382500   1.474927   1.427271
#>   DF.L2.C121 DF.L2.C122 DF.L2.C123 DF.L2.C124 DF.L2.C125 DF.L2.C126 DF.L2.C127
#> 1 0.04000991  0.9650389  0.0000000 0.55569038  1.0719146  0.4864181  0.9591814
#> 2 0.00000000  0.8305034  0.3839211 0.03766542  0.6090139  0.0000000  1.5819274
#> 3 0.00000000  0.2681344  0.0000000 0.18321518  0.5982721  0.0000000  1.2055265
#> 4 0.00000000  1.0785871  0.2089800 0.00000000  0.5867337  0.2937510  0.9878142
#> 5 0.04858260  0.5365197  0.0000000 0.39772500  0.8734708  0.2267082  1.0413327
#>   DF.L2.C128 DF.L2.C129 DF.L2.C130 DF.L2.C131 DF.L2.C132 DF.L2.C133 DF.L2.C134
#> 1   1.589170  0.8681959  0.6738689  0.7409457  0.8467491  0.7191947          0
#> 2   2.137798  1.1585408  0.5980641  1.0401187  0.8761826  0.7853955          0
#> 3   1.735495  1.1860956  0.4546392  1.1597105  0.8480308  0.8560016          0
#> 4   2.015504  0.5883152  0.5774118  0.6853674  0.7056450  0.8180591          0
#> 5   1.117168  0.9264638  0.2337522  0.8186213  1.0116111  0.4994296          0
#>   DF.L2.C135 DF.L2.C136 DF.L2.C137 DF.L2.C138 DF.L2.C139 DF.L2.C140 DF.L2.C141
#> 1 0.01551331  0.7999983          0 0.00000000  0.6119802 0.41804144  0.6473480
#> 2 0.00000000  0.0000000          0 0.00000000  0.3275911 0.09232707  0.4336147
#> 3 0.00000000  0.1432311          0 0.00000000  0.8778765 0.31251948  0.1740508
#> 4 0.03814402  0.6139485          0 0.38053896  0.0000000 0.66368898  0.9073214
#> 5 0.00000000  0.4978019          0 0.08365547  0.7636168 0.97733419  0.6743715
#>   DF.L2.C142 DF.L2.C143 DF.L2.C144 DF.L2.C145 DF.L2.C146 DF.L2.C147 DF.L2.C148
#> 1 0.09252497   1.384609          0  0.4995222  0.6374208   1.595091   1.409916
#> 2 0.00000000   1.579129          0  1.0792776  0.2188885   1.713568   2.088667
#> 3 0.00000000   1.385390          0  0.6410869  0.1960490   1.470011   1.871578
#> 4 0.33572657   1.350865          0  0.9411327  0.6221785   1.778196   1.662385
#> 5 0.25513777   1.127872          0  0.7732047  0.1651853   1.173136   1.498933
#>   DF.L2.C149 DF.L2.C150 DF.L2.C151 DF.L2.C152 DF.L2.C153 DF.L2.C154 DF.L2.C155
#> 1  1.3251904  1.4437667 0.22007613          0  0.7206724  0.1473649 0.62539907
#> 2  0.4039734  0.9450332 0.08839955          0  0.7528462  0.0000000 0.09786429
#> 3  0.6468219  1.2139183 0.20696327          0  0.4179446  0.0000000 0.48233625
#> 4  0.7907074  1.5775982 0.49180937          0  0.8814380  0.0000000 0.17950788
#> 5  1.2483646  1.4483450 0.89703818          0  0.8320260  0.2247276 0.80053111
#>   DF.L2.C156 DF.L2.C157 DF.L2.C158 DF.L2.C159 DF.L2.C160 DF.L2.C161 DF.L2.C162
#> 1 0.00000000  0.6802131  0.7480915  1.1044721  0.2854632  0.5299116  0.9757688
#> 2 0.00000000  0.3921681  0.8048391  1.1647706  0.2284707  0.7332045  0.9166294
#> 3 0.00000000  0.5108906  0.9107093  1.3805148  0.0000000  0.7152489  1.1659673
#> 4 0.09449086  0.3260290  0.3664192  0.6626278  0.0000000  1.0671971  1.7257741
#> 5 0.01848242  0.8753170  0.4826059  1.4583657  0.0000000  0.6230328  1.4901472
#>   DF.L2.C163 DF.L2.C164 DF.L2.C165 DF.L2.C166 DF.L2.C167 DF.L2.C168 DF.L2.C169
#> 1  0.7193666 0.07951702  0.5397253  1.3585101  0.7161059  0.0000000  1.2297138
#> 2  1.0256487 0.56224485  0.0000000  0.9328757  0.8608254  0.3512314  1.4312014
#> 3  0.7928005 0.39593247  0.3105304  0.9310840  0.7820088  0.0000000  1.1249391
#> 4  0.7232553 0.63144041  0.0000000  0.8466304  0.3918985  0.2204484  0.9004548
#> 5  0.1917715 0.36793638  0.5947824  1.6752705  1.0312023  0.0000000  0.8734395
#>   DF.L2.C170 DF.L2.C171 DF.L2.C172 DF.L2.C173 DF.L2.C174 DF.L2.C175 DF.L2.C176
#> 1  0.4609784 0.09677062  1.1063321  0.5336503  0.9343137 0.00000000  0.7592749
#> 2  0.0000000 0.00000000  0.5356164  0.3200235  0.8420396 0.56378595  1.7104965
#> 3  0.2099967 0.00000000  0.4901691  0.4416757  1.0025389 0.14725436  1.5451921
#> 4  0.0000000 0.00000000  0.7396949  1.2833503  1.1434589 0.03098376  1.6326630
#> 5  0.5952002 0.00000000  0.6787170  0.9238400  0.5246105 0.36796113  1.0654561
#>   DF.L2.C177 DF.L2.C178 DF.L2.C179 DF.L2.C180 DF.L2.C181 DF.L2.C182 DF.L2.C183
#> 1          0 0.12494723  0.4691063  0.8933488  1.4418640  0.9232943  0.8272396
#> 2          0 0.00794843  0.4606840  0.8968178  1.1435040  0.5905229  0.6185735
#> 3          0 0.00000000  0.3785866  0.8823257  1.3397827  0.6342214  0.9607753
#> 4          0 0.62504760  0.2070820  1.2152026  0.9786941  0.9544480  0.0168059
#> 5          0 0.59126823  0.0000000  1.2250733  0.8141017  0.8369020  0.5609418
#>   DF.L2.C184 DF.L2.C185 DF.L2.C186 DF.L2.C187 DF.L2.C188 DF.L2.C189 DF.L2.C190
#> 1  0.4089276  0.6682770  0.3699348  1.2492654  0.8397757  0.1521640  0.8482596
#> 2  0.0000000  0.4350805  0.5744467  1.9463881  0.7025375  0.0000000  0.6218581
#> 3  0.0000000  0.3500270  0.2504416  1.5130690  0.6066685  0.1082650  0.4383558
#> 4  0.0000000  0.5384196  0.6321931  1.4324989  0.9863478  0.0000000  0.8993088
#> 5  0.5523153  0.3804873  0.1985409  0.7484455  0.7215292  0.2193061  0.7202561
#>   DF.L2.C191 DF.L2.C192 DF.L2.C193 DF.L2.C194 DF.L2.C195 DF.L2.C196 DF.L2.C197
#> 1  0.8916476   1.392393  1.1485083  0.6921437 0.54896203  0.3679006  0.6847391
#> 2  0.6822568   1.677246  0.8587110  0.0000000 0.48345894  0.0000000  0.7217473
#> 3  0.7613560   1.543563  0.9320154  0.2351027 0.30591563  0.3714446  0.6958370
#> 4  0.1656999   1.803682  1.1192297  1.2236538 0.40720940  0.0000000  0.8710348
#> 5  0.5674340   1.574814  1.0017423  1.1137634 0.06168202  0.1938766  0.4969623
#>   DF.L2.C198 DF.L2.C199 DF.L2.C200
#> 1   1.330498   1.408009  1.0223246
#> 2   1.259764   1.152640  0.3751757
#> 3   1.371421   1.356450  0.3310389
#> 4   1.149764   1.134072  0.9822118
#> 5   1.335855   1.124828  0.7034305
#>  [ reached getOption("max.print") -- omitted 1 row ]

#if (h2o.deepwater.available()) {
#  prostate.dl = h2o.deepwater(x = 3:9, y = 2, backend="mxnet", training_frame = prostate.hex,
#                              hidden = c(100, 200), epochs = 5)
#  prostate.deepfeatures_layer1 =
#    h2o.deepfeatures(prostate.dl, prostate.hex, layer = "fc1_w")
#  prostate.deepfeatures_layer2 =
#    h2o.deepfeatures(prostate.dl, prostate.hex, layer = "fc2_w")
#  head(prostate.deepfeatures_layer1)
#  head(prostate.deepfeatures_layer2)
#}