Feature Generation via H2O Deep Learning or DeepWater Model

Extract the non-linear feature from an H2O data set using an H2O deep learning model.

h2o.deepfeatures(object, data, layer)

Arguments

object
An H2OModel object that represents the deep learning model to be used for feature extraction.

data
An H2OFrame object.

layer
Index (for DeepLearning, integer) or Name (for DeepWater, String) of the hidden layer to extract

Value

Returns an H2OFrame object with as many features as the number of units in the hidden layer of the specified index.

See also

link{h2o.deeplearning} for making H2O Deep Learning models. link{h2o.deepwater} for making H2O DeepWater models.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

prosPath = system.file("extdata", "prostate.csv", package = "h2o")
prostate.hex = h2o.importFile(path = prosPath)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

prostate.dl = h2o.deeplearning(x = 3:9, y = 2, training_frame = prostate.hex,
                               hidden = c(100, 200), epochs = 5)

#> Error in is.H2OFrame(training_frame): object 'prostate.hex' not found

prostate.deepfeatures_layer1 = h2o.deepfeatures(prostate.dl, prostate.hex, layer = 1)

#> Error in paste0("Predictions/models/", object@model_id, "/frames/", h2o.getId(data)): object 'prostate.dl' not found

prostate.deepfeatures_layer2 = h2o.deepfeatures(prostate.dl, prostate.hex, layer = 2)

#> Error in paste0("Predictions/models/", object@model_id, "/frames/", h2o.getId(data)): object 'prostate.dl' not found

head(prostate.deepfeatures_layer1)

#> Error in head(prostate.deepfeatures_layer1): object 'prostate.deepfeatures_layer1' not found

head(prostate.deepfeatures_layer2)

#> Error in head(prostate.deepfeatures_layer2): object 'prostate.deepfeatures_layer2' not found

#if (h2o.deepwater.available()) {
#  prostate.dl = h2o.deepwater(x = 3:9, y = 2, backend="mxnet", training_frame = prostate.hex,
#                              hidden = c(100, 200), epochs = 5)
#  prostate.deepfeatures_layer1 =
#    h2o.deepfeatures(prostate.dl, prostate.hex, layer = "fc1_w")
#  prostate.deepfeatures_layer2 =
#    h2o.deepfeatures(prostate.dl, prostate.hex, layer = "fc2_w")
#  head(prostate.deepfeatures_layer1)
#  head(prostate.deepfeatures_layer2)
#}