Get a grid object from H2O distributed K/V store.

Note that if neither cross-validation nor a validation frame is used in the grid search, then the training metrics will display in the "get grid" output. If a validation frame is passed to the grid, and nfolds = 0, then the validation metrics will display. However, if nfolds > 1, then cross-validation metrics will display even if a validation frame is provided.

h2o.getGrid(grid_id, sort_by, decreasing)

Arguments

grid_id
ID of existing grid object to fetch

sort_by
Sort the models in the grid space by a metric. Choices are "logloss", "residual_deviance", "mse", "auc", "accuracy", "precision", "recall", "f1", etc.

decreasing
Specify whether sort order should be decreasing

Examples

library(h2o)
library(jsonlite)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

iris.hex <- as.h2o(iris)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

h2o.grid("gbm", grid_id = "gbm_grid_id", x = c(1:4), y = 5,
         training_frame = iris.hex, hyper_params = list(ntrees = c(1,2,3)))

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Hyper-parameter: ntrees, 1
#> [2017-12-11 20:07:42] failure_details: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y 
#> [2017-12-11 20:07:42] failure_stack_traces: java.lang.RuntimeException: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> 	at water.fvec.RollupStats.get(RollupStats.java:322)
#> 	at water.fvec.RollupStats.get(RollupStats.java:352)
#> 	at water.fvec.Vec.rollupStats(Vec.java:847)
#> 	at water.fvec.Vec.checksum_impl(Vec.java:866)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at water.fvec.Frame.checksum_impl(Frame.java:558)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at hex.Model$Parameters.checksum_impl(Model.java:408)
#> 	at hex.Model$Parameters.checksum(Model.java:339)
#> 	at hex.grid.GridSearch$2.filter(GridSearch.java:298)
#> 	at water.KeySnapshot.filter(KeySnapshot.java:65)
#> 	at hex.grid.GridSearch.buildModel(GridSearch.java:290)
#> 	at hex.grid.GridSearch.gridSearch(GridSearch.java:213)
#> 	at hex.grid.GridSearch.access$000(GridSearch.java:68)
#> 	at hex.grid.GridSearch$1.compute2(GridSearch.java:135)
#> 	at water.H2O$H2OCountedCompleter.compute(H2O.java:1263)
#> 	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
#> 	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
#> 	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:974)
#> 	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1477)
#> 	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
#>  
#> Hyper-parameter: ntrees, 2
#> [2017-12-11 20:07:42] failure_details: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y 
#> [2017-12-11 20:07:42] failure_stack_traces: java.lang.RuntimeException: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> 	at water.fvec.RollupStats.get(RollupStats.java:322)
#> 	at water.fvec.RollupStats.get(RollupStats.java:352)
#> 	at water.fvec.Vec.rollupStats(Vec.java:847)
#> 	at water.fvec.Vec.checksum_impl(Vec.java:866)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at water.fvec.Frame.checksum_impl(Frame.java:558)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at hex.Model$Parameters.checksum_impl(Model.java:408)
#> 	at hex.Model$Parameters.checksum(Model.java:339)
#> 	at hex.grid.GridSearch$2.filter(GridSearch.java:298)
#> 	at water.KeySnapshot.filter(KeySnapshot.java:65)
#> 	at hex.grid.GridSearch.buildModel(GridSearch.java:290)
#> 	at hex.grid.GridSearch.gridSearch(GridSearch.java:213)
#> 	at hex.grid.GridSearch.access$000(GridSearch.java:68)
#> 	at hex.grid.GridSearch$1.compute2(GridSearch.java:135)
#> 	at water.H2O$H2OCountedCompleter.compute(H2O.java:1263)
#> 	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
#> 	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
#> 	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:974)
#> 	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1477)
#> 	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
#>  
#> Hyper-parameter: ntrees, 3
#> [2017-12-11 20:07:42] failure_details: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y 
#> [2017-12-11 20:07:42] failure_stack_traces: java.lang.RuntimeException: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> 	at water.fvec.RollupStats.get(RollupStats.java:322)
#> 	at water.fvec.RollupStats.get(RollupStats.java:352)
#> 	at water.fvec.Vec.rollupStats(Vec.java:847)
#> 	at water.fvec.Vec.checksum_impl(Vec.java:866)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at water.fvec.Frame.checksum_impl(Frame.java:558)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at hex.Model$Parameters.checksum_impl(Model.java:408)
#> 	at hex.Model$Parameters.checksum(Model.java:339)
#> 	at hex.grid.GridSearch$2.filter(GridSearch.java:298)
#> 	at water.KeySnapshot.filter(KeySnapshot.java:65)
#> 	at hex.grid.GridSearch.buildModel(GridSearch.java:290)
#> 	at hex.grid.GridSearch.gridSearch(GridSearch.java:213)
#> 	at hex.grid.GridSearch.access$000(GridSearch.java:68)
#> 	at hex.grid.GridSearch$1.compute2(GridSearch.java:135)
#> 	at water.H2O$H2OCountedCompleter.compute(H2O.java:1263)
#> 	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
#> 	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
#> 	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:974)
#> 	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1477)
#> 	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
#>  

#> H2O Grid Details
#> ================
#> 
#> Grid ID: gbm_grid_id 
#> Used hyper parameters: 
#>   -  ntrees 
#> Number of models: 0 
#> Number of failed models: 3 
#> 
#> NULL
#> Failed models
#> -------------
#>  ntrees status_failed
#>       1          FAIL
#>       2          FAIL
#>       3          FAIL
#>                                                                                           msgs_failed
#>  "Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y"
#>  "Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y"
#>  "Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y"

grid <- h2o.getGrid("gbm_grid_id")

#> Hyper-parameter: ntrees, 1
#> [2017-12-11 20:07:42] failure_details: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y 
#> [2017-12-11 20:07:42] failure_stack_traces: java.lang.RuntimeException: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> 	at water.fvec.RollupStats.get(RollupStats.java:322)
#> 	at water.fvec.RollupStats.get(RollupStats.java:352)
#> 	at water.fvec.Vec.rollupStats(Vec.java:847)
#> 	at water.fvec.Vec.checksum_impl(Vec.java:866)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at water.fvec.Frame.checksum_impl(Frame.java:558)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at hex.Model$Parameters.checksum_impl(Model.java:408)
#> 	at hex.Model$Parameters.checksum(Model.java:339)
#> 	at hex.grid.GridSearch$2.filter(GridSearch.java:298)
#> 	at water.KeySnapshot.filter(KeySnapshot.java:65)
#> 	at hex.grid.GridSearch.buildModel(GridSearch.java:290)
#> 	at hex.grid.GridSearch.gridSearch(GridSearch.java:213)
#> 	at hex.grid.GridSearch.access$000(GridSearch.java:68)
#> 	at hex.grid.GridSearch$1.compute2(GridSearch.java:135)
#> 	at water.H2O$H2OCountedCompleter.compute(H2O.java:1263)
#> 	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
#> 	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
#> 	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:974)
#> 	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1477)
#> 	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
#>  
#> Hyper-parameter: ntrees, 2
#> [2017-12-11 20:07:42] failure_details: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y 
#> [2017-12-11 20:07:42] failure_stack_traces: java.lang.RuntimeException: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> 	at water.fvec.RollupStats.get(RollupStats.java:322)
#> 	at water.fvec.RollupStats.get(RollupStats.java:352)
#> 	at water.fvec.Vec.rollupStats(Vec.java:847)
#> 	at water.fvec.Vec.checksum_impl(Vec.java:866)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at water.fvec.Frame.checksum_impl(Frame.java:558)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at hex.Model$Parameters.checksum_impl(Model.java:408)
#> 	at hex.Model$Parameters.checksum(Model.java:339)
#> 	at hex.grid.GridSearch$2.filter(GridSearch.java:298)
#> 	at water.KeySnapshot.filter(KeySnapshot.java:65)
#> 	at hex.grid.GridSearch.buildModel(GridSearch.java:290)
#> 	at hex.grid.GridSearch.gridSearch(GridSearch.java:213)
#> 	at hex.grid.GridSearch.access$000(GridSearch.java:68)
#> 	at hex.grid.GridSearch$1.compute2(GridSearch.java:135)
#> 	at water.H2O$H2OCountedCompleter.compute(H2O.java:1263)
#> 	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
#> 	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
#> 	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:974)
#> 	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1477)
#> 	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
#>  
#> Hyper-parameter: ntrees, 3
#> [2017-12-11 20:07:42] failure_details: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y 
#> [2017-12-11 20:07:42] failure_stack_traces: java.lang.RuntimeException: Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> 	at water.fvec.RollupStats.get(RollupStats.java:322)
#> 	at water.fvec.RollupStats.get(RollupStats.java:352)
#> 	at water.fvec.Vec.rollupStats(Vec.java:847)
#> 	at water.fvec.Vec.checksum_impl(Vec.java:866)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at water.fvec.Frame.checksum_impl(Frame.java:558)
#> 	at water.Keyed.checksum(Keyed.java:69)
#> 	at hex.Model$Parameters.checksum_impl(Model.java:408)
#> 	at hex.Model$Parameters.checksum(Model.java:339)
#> 	at hex.grid.GridSearch$2.filter(GridSearch.java:298)
#> 	at water.KeySnapshot.filter(KeySnapshot.java:65)
#> 	at hex.grid.GridSearch.buildModel(GridSearch.java:290)
#> 	at hex.grid.GridSearch.gridSearch(GridSearch.java:213)
#> 	at hex.grid.GridSearch.access$000(GridSearch.java:68)
#> 	at hex.grid.GridSearch$1.compute2(GridSearch.java:135)
#> 	at water.H2O$H2OCountedCompleter.compute(H2O.java:1263)
#> 	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
#> 	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
#> 	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:974)
#> 	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1477)
#> 	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
#>  

# Get grid summary
summary(grid)

#> H2O Grid Details
#> ================
#> 
#> Grid ID: gbm_grid_id 
#> Used hyper parameters: 
#>   -  ntrees 
#> Number of models: 0 
#> Number of failed models: 3 
#> 
#> NULL
#> Failed models
#> -------------
#>  ntrees status_failed
#>       1          FAIL
#>       2          FAIL
#>       3          FAIL
#>                                                                                           msgs_failed
#>  "Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y"
#>  "Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y"
#>  "Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y"
#> H2O Grid Summary
#> ================
#> 
#> Grid ID: gbm_grid_id 
#> Used hyper parameters: 
#>   -  ntrees 
#> Number of models: 0 
#> 
#> Number of failed models: 3 
#>   -  Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y  -  Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y  -  Rollups not possible, because Vec was deleted: $04ff19000000ffffffffc137e5cd560781aba041ddbae2${~Y
#> Note: To see exception stack traces please pass parameter `show_stack_traces = T` to this function.

# Fetch grid models
model_ids <- grid@model_ids
models <- lapply(model_ids, function(id) { h2o.getModel(id)})