Get a grid object from H2O distributed K/V store.

Note that if neither cross-validation nor a validation frame is used in the grid search, then the training metrics will display in the "get grid" output. If a validation frame is passed to the grid, and nfolds = 0, then the validation metrics will display. However, if nfolds > 1, then cross-validation metrics will display even if a validation frame is provided.

h2o.getGrid(grid_id, sort_by, decreasing)

Arguments

grid_id
ID of existing grid object to fetch

sort_by
Sort the models in the grid space by a metric. Choices are "logloss", "residual_deviance", "mse", "auc", "accuracy", "precision", "recall", "f1", etc.

decreasing
Specify whether sort order should be decreasing

Examples

library(h2o)
library(jsonlite)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

iris.hex <- as.h2o(iris)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

h2o.grid("gbm", grid_id = "gbm_grid_id", x = c(1:4), y = 5,
         training_frame = iris.hex, hyper_params = list(ntrees = c(1,2,3)))

#> Error in h2o.grid("gbm", grid_id = "gbm_grid_id", x = c(1:4), y = 5, training_frame = iris.hex,     hyper_params = list(ntrees = c(1, 2, 3))): object 'iris.hex' not found

grid <- h2o.getGrid("gbm_grid_id")

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

# Get grid summary
summary(grid)

#> Error in object[[i]]: object of type 'closure' is not subsettable

# Fetch grid models
model_ids <- grid@model_ids

#> Error in eval(expr, envir, enclos): trying to get slot "model_ids" from an object of a basic class ("function") with no slots

models <- lapply(model_ids, function(id) { h2o.getModel(id)})

#> Error in lapply(model_ids, function(id) {    h2o.getModel(id)}): object 'model_ids' not found