Categorical Interaction Feature Creation in H2O

Creates a data frame in H2O with n-th order interaction features between categorical columns, as specified by the user.

h2o.interaction(data, destination_frame, factors, pairwise, max_factors,
  min_occurrence)

Arguments

data
An H2OFrame object containing the categorical columns.

destination_frame
A string indicating the destination key. If empty, this will be auto-generated by H2O.

factors
Factor columns (either indices or column names).

pairwise
Whether to create pairwise interactions between factors (otherwise create one higher-order interaction). Only applicable if there are 3 or more factors.

max_factors
Max. number of factor levels in pair-wise interaction terms (if enforced, one extra catch-all factor will be made)

min_occurrence
Min. occurrence threshold for factor levels in pair-wise interaction terms

Value

Returns an H2OFrame object.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

# Create some random data
myframe <- h2o.createFrame(rows = 20, cols = 5,
                         seed = -12301283, randomize = TRUE, value = 0,
                         categorical_fraction = 0.8, factors = 10, real_range = 1,
                         integer_fraction = 0.2, integer_range = 10,
                         binary_fraction = 0, binary_ones_fraction = 0.5,
                         missing_fraction = 0.2,
                         response_factors = 1)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

# Turn integer column into a categorical
myframe[,5] <- as.factor(myframe[,5])
head(myframe, 20)

#>       C1    C2    C3    C4   C5
#> 1  c0.l7 c1.l3 c2.l7 c3.l2    8
#> 2  c0.l7 c1.l7 c2.l2 c3.l3    3
#> 3  c0.l5 c1.l1 c2.l6 c3.l8   -4
#> 4   <NA>  <NA>  <NA>  <NA>  -10
#> 5  c0.l1 c1.l5  <NA> c3.l9    7
#> 6  c0.l9 c1.l7 c2.l5 c3.l7   -4
#> 7  c0.l9 c1.l9 c2.l2  <NA> <NA>
#> 8  c0.l0 c1.l5 c2.l4  <NA>    2
#> 9  c0.l3 c1.l6 c2.l6 c3.l6    3
#> 10 c0.l1 c1.l2 c2.l3 c3.l9   -5
#> 11 c0.l8 c1.l4  <NA>  <NA>   -1
#> 12  <NA> c1.l2  <NA> c3.l4   -1
#> 13 c0.l8  <NA> c2.l1 c3.l7   -7
#> 14 c0.l5 c1.l5 c2.l1 c3.l3 <NA>
#> 15 c0.l8 c1.l2 c2.l1 c3.l8 <NA>
#> 16 c0.l6 c1.l1 c2.l2 c3.l3    7
#> 17 c0.l3 c1.l8 c2.l9 c3.l8   -6
#> 18 c0.l9 c1.l1 c2.l7 c3.l9   -9
#> 19 c0.l1 c1.l9  <NA> c3.l7 <NA>
#> 20 c0.l3 c1.l4 c2.l5 c3.l3   10

# Create pairwise interactions
pairwise <- h2o.interaction(myframe, destination_frame = 'pairwise',
                            factors = list(c(1,2),c("C2","C3","C4")),
                            pairwise=TRUE, max_factors = 10, min_occurrence = 1)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

head(pairwise, 20)

#>          C1_C2       C2_C3       C2_C4       C3_C4
#> 1  c0.l7_c1.l3 c1.l3_c2.l7 c1.l3_c3.l2 c2.l7_c3.l2
#> 2  c0.l7_c1.l7 c1.l7_c2.l2 c1.l7_c3.l3 c2.l2_c3.l3
#> 3  c0.l5_c1.l1 c1.l1_c2.l6 c1.l1_c3.l8       other
#> 4        NA_NA       NA_NA       NA_NA       NA_NA
#> 5  c0.l1_c1.l5       other       other       other
#> 6        other c1.l7_c2.l5 c1.l7_c3.l7 c2.l5_c3.l7
#> 7  c0.l9_c1.l9       other       other       other
#> 8  c0.l0_c1.l5 c1.l5_c2.l4       other       other
#> 9  c0.l3_c1.l6 c1.l6_c2.l6 c1.l6_c3.l6 c2.l6_c3.l6
#> 10 c0.l1_c1.l2 c1.l2_c2.l3       other       other
#> 11       other       other       other       NA_NA
#> 12       other       other c1.l2_c3.l4       other
#> 13       other       other       other c2.l1_c3.l7
#> 14 c0.l5_c1.l5 c1.l5_c2.l1 c1.l5_c3.l3 c2.l1_c3.l3
#> 15       other c1.l2_c2.l1       other c2.l1_c3.l8
#> 16       other       other c1.l1_c3.l3 c2.l2_c3.l3
#> 17       other       other c1.l8_c3.l8 c2.l9_c3.l8
#> 18       other       other       other       other
#> 19       other       other       other       other
#> 20       other       other       other c2.l5_c3.l3

h2o.levels(pairwise,2)

#>  [1] "NA_NA"       "c1.l5_c2.l4" "c1.l7_c2.l5" "c1.l6_c2.l6" "c1.l3_c2.l7"
#>  [6] "c1.l7_c2.l2" "c1.l2_c2.l3" "c1.l1_c2.l6" "c1.l5_c2.l1" "c1.l2_c2.l1"
#> [11] "other"      

# Create 5-th order interaction
higherorder <- h2o.interaction(myframe, destination_frame = 'higherorder', factors = c(1,2,3,4,5),
                               pairwise=FALSE, max_factors = 10000, min_occurrence = 1)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

head(higherorder, 20)

#>                C1_C2_C3_C4_C5
#> 1   c0.l7_c1.l3_c2.l7_c3.l2_8
#> 2   c0.l7_c1.l7_c2.l2_c3.l3_3
#> 3  c0.l5_c1.l1_c2.l6_c3.l8_-4
#> 4             NA_NA_NA_NA_-10
#> 5      c0.l1_c1.l5_NA_c3.l9_7
#> 6  c0.l9_c1.l7_c2.l5_c3.l7_-4
#> 7     c0.l9_c1.l9_c2.l2_NA_NA
#> 8      c0.l0_c1.l5_c2.l4_NA_2
#> 9   c0.l3_c1.l6_c2.l6_c3.l6_3
#> 10 c0.l1_c1.l2_c2.l3_c3.l9_-5
#> 11       c0.l8_c1.l4_NA_NA_-1
#> 12       NA_c1.l2_NA_c3.l4_-1
#> 13    c0.l8_NA_c2.l1_c3.l7_-7
#> 14 c0.l5_c1.l5_c2.l1_c3.l3_NA
#> 15 c0.l8_c1.l2_c2.l1_c3.l8_NA
#> 16  c0.l6_c1.l1_c2.l2_c3.l3_7
#> 17 c0.l3_c1.l8_c2.l9_c3.l8_-6
#> 18 c0.l9_c1.l1_c2.l7_c3.l9_-9
#> 19    c0.l1_c1.l9_NA_c3.l7_NA
#> 20 c0.l3_c1.l4_c2.l5_c3.l3_10

# Limit the number of factors of the "categoricalized" integer column
# to at most 3 factors, and only if they occur at least twice
head(myframe[,5], 20)

#>      C5
#> 1     8
#> 2     3
#> 3    -4
#> 4   -10
#> 5     7
#> 6    -4
#> 7  <NA>
#> 8     2
#> 9     3
#> 10   -5
#> 11   -1
#> 12   -1
#> 13   -7
#> 14 <NA>
#> 15 <NA>
#> 16    7
#> 17   -6
#> 18   -9
#> 19 <NA>
#> 20   10

trim_integer_levels <- h2o.interaction(myframe, destination_frame = 'trim_integers', factors = "C5",
                                       pairwise = FALSE, max_factors = 3, min_occurrence = 2)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

head(trim_integer_levels, 20)

#>    C5_C5
#> 1  other
#> 2      3
#> 3     -4
#> 4  other
#> 5  other
#> 6     -4
#> 7   <NA>
#> 8  other
#> 9      3
#> 10 other
#> 11    -1
#> 12    -1
#> 13 other
#> 14  <NA>
#> 15  <NA>
#> 16 other
#> 17 other
#> 18 other
#> 19  <NA>
#> 20 other

# Put all together
myframe <- h2o.cbind(myframe, pairwise, higherorder, trim_integer_levels)
myframe

#>      C1    C2    C3    C4  C5       C1_C2       C2_C3       C2_C4       C3_C4
#> 1 c0.l7 c1.l3 c2.l7 c3.l2   8 c0.l7_c1.l3 c1.l3_c2.l7 c1.l3_c3.l2 c2.l7_c3.l2
#> 2 c0.l7 c1.l7 c2.l2 c3.l3   3 c0.l7_c1.l7 c1.l7_c2.l2 c1.l7_c3.l3 c2.l2_c3.l3
#> 3 c0.l5 c1.l1 c2.l6 c3.l8  -4 c0.l5_c1.l1 c1.l1_c2.l6 c1.l1_c3.l8       other
#> 4  <NA>  <NA>  <NA>  <NA> -10       NA_NA       NA_NA       NA_NA       NA_NA
#> 5 c0.l1 c1.l5  <NA> c3.l9   7 c0.l1_c1.l5       other       other       other
#> 6 c0.l9 c1.l7 c2.l5 c3.l7  -4       other c1.l7_c2.l5 c1.l7_c3.l7 c2.l5_c3.l7
#>               C1_C2_C3_C4_C5 C5_C5
#> 1  c0.l7_c1.l3_c2.l7_c3.l2_8 other
#> 2  c0.l7_c1.l7_c2.l2_c3.l3_3     3
#> 3 c0.l5_c1.l1_c2.l6_c3.l8_-4    -4
#> 4            NA_NA_NA_NA_-10 other
#> 5     c0.l1_c1.l5_NA_c3.l9_7 other
#> 6 c0.l9_c1.l7_c2.l5_c3.l7_-4    -4
#> 
#> [20 rows x 11 columns] 

head(myframe,20)

#>       C1    C2    C3    C4   C5       C1_C2       C2_C3       C2_C4       C3_C4
#> 1  c0.l7 c1.l3 c2.l7 c3.l2    8 c0.l7_c1.l3 c1.l3_c2.l7 c1.l3_c3.l2 c2.l7_c3.l2
#> 2  c0.l7 c1.l7 c2.l2 c3.l3    3 c0.l7_c1.l7 c1.l7_c2.l2 c1.l7_c3.l3 c2.l2_c3.l3
#> 3  c0.l5 c1.l1 c2.l6 c3.l8   -4 c0.l5_c1.l1 c1.l1_c2.l6 c1.l1_c3.l8       other
#> 4   <NA>  <NA>  <NA>  <NA>  -10       NA_NA       NA_NA       NA_NA       NA_NA
#> 5  c0.l1 c1.l5  <NA> c3.l9    7 c0.l1_c1.l5       other       other       other
#> 6  c0.l9 c1.l7 c2.l5 c3.l7   -4       other c1.l7_c2.l5 c1.l7_c3.l7 c2.l5_c3.l7
#> 7  c0.l9 c1.l9 c2.l2  <NA> <NA> c0.l9_c1.l9       other       other       other
#> 8  c0.l0 c1.l5 c2.l4  <NA>    2 c0.l0_c1.l5 c1.l5_c2.l4       other       other
#> 9  c0.l3 c1.l6 c2.l6 c3.l6    3 c0.l3_c1.l6 c1.l6_c2.l6 c1.l6_c3.l6 c2.l6_c3.l6
#> 10 c0.l1 c1.l2 c2.l3 c3.l9   -5 c0.l1_c1.l2 c1.l2_c2.l3       other       other
#> 11 c0.l8 c1.l4  <NA>  <NA>   -1       other       other       other       NA_NA
#> 12  <NA> c1.l2  <NA> c3.l4   -1       other       other c1.l2_c3.l4       other
#> 13 c0.l8  <NA> c2.l1 c3.l7   -7       other       other       other c2.l1_c3.l7
#> 14 c0.l5 c1.l5 c2.l1 c3.l3 <NA> c0.l5_c1.l5 c1.l5_c2.l1 c1.l5_c3.l3 c2.l1_c3.l3
#> 15 c0.l8 c1.l2 c2.l1 c3.l8 <NA>       other c1.l2_c2.l1       other c2.l1_c3.l8
#> 16 c0.l6 c1.l1 c2.l2 c3.l3    7       other       other c1.l1_c3.l3 c2.l2_c3.l3
#> 17 c0.l3 c1.l8 c2.l9 c3.l8   -6       other       other c1.l8_c3.l8 c2.l9_c3.l8
#> 18 c0.l9 c1.l1 c2.l7 c3.l9   -9       other       other       other       other
#> 19 c0.l1 c1.l9  <NA> c3.l7 <NA>       other       other       other       other
#> 20 c0.l3 c1.l4 c2.l5 c3.l3   10       other       other       other c2.l5_c3.l3
#>                C1_C2_C3_C4_C5 C5_C5
#> 1   c0.l7_c1.l3_c2.l7_c3.l2_8 other
#> 2   c0.l7_c1.l7_c2.l2_c3.l3_3     3
#> 3  c0.l5_c1.l1_c2.l6_c3.l8_-4    -4
#> 4             NA_NA_NA_NA_-10 other
#> 5      c0.l1_c1.l5_NA_c3.l9_7 other
#> 6  c0.l9_c1.l7_c2.l5_c3.l7_-4    -4
#> 7     c0.l9_c1.l9_c2.l2_NA_NA  <NA>
#> 8      c0.l0_c1.l5_c2.l4_NA_2 other
#> 9   c0.l3_c1.l6_c2.l6_c3.l6_3     3
#> 10 c0.l1_c1.l2_c2.l3_c3.l9_-5 other
#> 11       c0.l8_c1.l4_NA_NA_-1    -1
#> 12       NA_c1.l2_NA_c3.l4_-1    -1
#> 13    c0.l8_NA_c2.l1_c3.l7_-7 other
#> 14 c0.l5_c1.l5_c2.l1_c3.l3_NA  <NA>
#> 15 c0.l8_c1.l2_c2.l1_c3.l8_NA  <NA>
#> 16  c0.l6_c1.l1_c2.l2_c3.l3_7 other
#> 17 c0.l3_c1.l8_c2.l9_c3.l8_-6 other
#> 18 c0.l9_c1.l1_c2.l7_c3.l9_-9 other
#> 19    c0.l1_c1.l9_NA_c3.l7_NA  <NA>
#> 20 c0.l3_c1.l4_c2.l5_c3.l3_10 other

summary(myframe)

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  C1       C2       C3       C4       C5     C1_C2           C2_C3          
#>  c0.l1:3  c1.l1:3  c2.l1:3  c3.l3:4  -4 :2  other      :10  other      :10 
#>  c0.l3:3  c1.l2:3  c2.l2:3  c3.l7:3  -1 :2  c0.l7_c1.l7: 1  NA_NA      : 1 
#>  c0.l8:3  c1.l5:3  c2.l5:2  c3.l8:3  3  :2  NA_NA      : 1  c1.l5_c2.l4: 1 
#>  c0.l9:3  c1.l4:2  c2.l6:2  c3.l9:3  7  :2  c0.l9_c1.l9: 1  c1.l7_c2.l5: 1 
#>  c0.l5:2  c1.l7:2  c2.l7:2  c3.l2:1  -10:1  c0.l1_c1.l2: 1  c1.l6_c2.l6: 1 
#>  c0.l7:2  c1.l9:2  c2.l3:1  c3.l4:1  -9 :1  c0.l7_c1.l3: 1  c1.l3_c2.l7: 1 
#>  NA   :2  NA   :2  NA   :5  NA   :4  NA :4                                 
#>  C2_C4           C3_C4          C1_C2_C3_C4_C5                C5_C5    
#>  other      :10  other      :8  c0.l6_c1.l1_c2.l2_c3.l3_7 :1  other:10 
#>  NA_NA      : 1  NA_NA      :2  c0.l1_c1.l2_c2.l3_c3.l9_-5:1  -4   : 2 
#>  c1.l3_c3.l2: 1  c2.l2_c3.l3:2  c0.l8_c1.l4_NA_NA_-1      :1  -1   : 2 
#>  c1.l7_c3.l7: 1  c2.l5_c3.l7:1  NA_c1.l2_NA_c3.l4_-1      :1  3    : 2 
#>  c1.l6_c3.l6: 1  c2.l6_c3.l6:1  c0.l9_c1.l1_c2.l7_c3.l9_-9:1  NA   : 4 
#>  c1.l7_c3.l3: 1  c2.l1_c3.l3:1  c0.l3_c1.l4_c2.l5_c3.l3_10:1           
#>