Model Performance Metrics in H2O

Given a trained h2o model, compute its performance on the given dataset

h2o.performance(model, newdata = NULL, train = FALSE, valid = FALSE,
  xval = FALSE, data = NULL)

Arguments

model
An H2OModel object

newdata
An H2OFrame. The model will make predictions on this dataset, and subsequently score them. The dataset should match the dataset that was used to train the model, in terms of column names, types, and dimensions. If newdata is passed in, then train, valid, and xval are ignored.

train
A logical value indicating whether to return the training metrics (constructed during training).

Note: when the trained h2o model uses balance_classes, the training metrics constructed during training will be from the balanced training dataset. For more information visit: https://0xdata.atlassian.net/browse/TN-9

valid
A logical value indicating whether to return the validation metrics (constructed during training).

xval
A logical value indicating whether to return the cross-validation metrics (constructed during training).

data
(DEPRECATED) An H2OFrame. This argument is now called `newdata`.

Value

Returns an object of the H2OModelMetrics subclass.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

prostate.hex$CAPSULE <- as.factor(prostate.hex$CAPSULE)
prostate.gbm <- h2o.gbm(3:9, "CAPSULE", prostate.hex)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================================================================| 100%

h2o.performance(model = prostate.gbm, newdata=prostate.hex)

#> H2OBinomialMetrics: gbm
#> 
#> MSE:  0.07338612
#> RMSE:  0.2708987
#> LogLoss:  0.2675724
#> Mean Per-Class Error:  0.07431401
#> AUC:  0.9801618
#> Gini:  0.9603236
#> 
#> Confusion Matrix (vertical: actual; across: predicted) for F1-optimal threshold:
#>          0   1    Error     Rate
#> 0      217  10 0.044053  =10/227
#> 1       16 137 0.104575  =16/153
#> Totals 233 147 0.068421  =26/380
#> 
#> Maximum Metrics: Maximum metrics at their respective thresholds
#>                         metric threshold    value idx
#> 1                       max f1  0.477235 0.913333 146
#> 2                       max f2  0.339414 0.947302 184
#> 3                 max f0point5  0.579116 0.934230 128
#> 4                 max accuracy  0.477235 0.931579 146
#> 5                max precision  0.982481 1.000000   0
#> 6                   max recall  0.189702 1.000000 233
#> 7              max specificity  0.982481 1.000000   0
#> 8             max absolute_mcc  0.477235 0.857317 146
#> 9   max min_per_class_accuracy  0.441872 0.908497 156
#> 10 max mean_per_class_accuracy  0.477235 0.925686 146
#> 
#> Gains/Lift Table: Extract with `h2o.gainsLift(<model>, <data>)` or `h2o.gainsLift(<model>, valid=<T/F>, xval=<T/F>)`

## If model uses balance_classes
## the results from train = TRUE will not match the results from newdata = prostate.hex
prostate.gbm.balanced <- h2o.gbm(3:9, "CAPSULE", prostate.hex, balance_classes = TRUE)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================                                                |  32%  |                                                                              |======================================================================| 100%

h2o.performance(model = prostate.gbm.balanced, newdata = prostate.hex)

#> H2OBinomialMetrics: gbm
#> 
#> MSE:  0.07613993
#> RMSE:  0.2759346
#> LogLoss:  0.269857
#> Mean Per-Class Error:  0.08752987
#> AUC:  0.974648
#> Gini:  0.949296
#> 
#> Confusion Matrix (vertical: actual; across: predicted) for F1-optimal threshold:
#>          0   1    Error     Rate
#> 0      211  16 0.070485  =16/227
#> 1       16 137 0.104575  =16/153
#> Totals 227 153 0.084211  =32/380
#> 
#> Maximum Metrics: Maximum metrics at their respective thresholds
#>                         metric threshold    value idx
#> 1                       max f1  0.446296 0.895425 152
#> 2                       max f2  0.300473 0.922208 182
#> 3                 max f0point5  0.567457 0.923191 130
#> 4                 max accuracy  0.446296 0.915789 152
#> 5                max precision  0.971155 1.000000   0
#> 6                   max recall  0.157209 1.000000 234
#> 7              max specificity  0.971155 1.000000   0
#> 8             max absolute_mcc  0.446296 0.824940 152
#> 9   max min_per_class_accuracy  0.396491 0.903084 160
#> 10 max mean_per_class_accuracy  0.446296 0.912470 152
#> 
#> Gains/Lift Table: Extract with `h2o.gainsLift(<model>, <data>)` or `h2o.gainsLift(<model>, valid=<T/F>, xval=<T/F>)`

h2o.performance(model = prostate.gbm.balanced, train = TRUE)

#> H2OBinomialMetrics: gbm
#> ** Reported on training data. **
#> 
#> MSE:  0.07583012
#> RMSE:  0.2753727
#> LogLoss:  0.2720462
#> Mean Per-Class Error:  0.07862129
#> AUC:  0.979442
#> Gini:  0.958884
#> 
#> Confusion Matrix (vertical: actual; across: predicted) for F1-optimal threshold:
#>          0   1    Error     Rate
#> 0      211  16 0.070485  =16/227
#> 1       19 200 0.086758  =19/219
#> Totals 230 216 0.078475  =35/446
#> 
#> Maximum Metrics: Maximum metrics at their respective thresholds
#>                         metric threshold    value idx
#> 1                       max f1  0.446296 0.919540 152
#> 2                       max f2  0.300473 0.944987 182
#> 3                 max f0point5  0.587449 0.942327 126
#> 4                 max accuracy  0.446296 0.921525 152
#> 5                max precision  0.971155 1.000000   0
#> 6                   max recall  0.157209 1.000000 234
#> 7              max specificity  0.971155 1.000000   0
#> 8             max absolute_mcc  0.446296 0.843037 152
#> 9   max min_per_class_accuracy  0.446296 0.913242 152
#> 10 max mean_per_class_accuracy  0.446296 0.921379 152
#> 
#> Gains/Lift Table: Extract with `h2o.gainsLift(<model>, <data>)` or `h2o.gainsLift(<model>, valid=<T/F>, xval=<T/F>)`