Model Performance Metrics in H2O

Given a trained h2o model, compute its performance on the given dataset

h2o.performance(model, newdata = NULL, train = FALSE, valid = FALSE,
  xval = FALSE, data = NULL)

Arguments

model
An H2OModel object

newdata
An H2OFrame. The model will make predictions on this dataset, and subsequently score them. The dataset should match the dataset that was used to train the model, in terms of column names, types, and dimensions. If newdata is passed in, then train, valid, and xval are ignored.

train
A logical value indicating whether to return the training metrics (constructed during training).

Note: when the trained h2o model uses balance_classes, the training metrics constructed during training will be from the balanced training dataset. For more information visit: https://0xdata.atlassian.net/browse/TN-9

valid
A logical value indicating whether to return the validation metrics (constructed during training).

xval
A logical value indicating whether to return the cross-validation metrics (constructed during training).

data
(DEPRECATED) An H2OFrame. This argument is now called `newdata`.

Value

Returns an object of the H2OModelMetrics subclass.

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

prostate.hex$CAPSULE <- as.factor(prostate.hex$CAPSULE)

#> Error in is.H2OFrame(x): object 'prostate.hex' not found

prostate.gbm <- h2o.gbm(3:9, "CAPSULE", prostate.hex)

#> Error in is.H2OFrame(training_frame): object 'prostate.hex' not found

h2o.performance(model = prostate.gbm, newdata=prostate.hex)

#> Error in is(model, "H2OModel"): object 'prostate.gbm' not found

## If model uses balance_classes
## the results from train = TRUE will not match the results from newdata = prostate.hex
prostate.gbm.balanced <- h2o.gbm(3:9, "CAPSULE", prostate.hex, balance_classes = TRUE)

#> Error in is.H2OFrame(training_frame): object 'prostate.hex' not found

h2o.performance(model = prostate.gbm.balanced, newdata = prostate.hex)

#> Error in is(model, "H2OModel"): object 'prostate.gbm.balanced' not found

h2o.performance(model = prostate.gbm.balanced, train = TRUE)

#> Error in is(model, "H2OModel"): object 'prostate.gbm.balanced' not found