Principal components analysis of an H2O data frame using the power method to calculate the singular value decomposition of the Gram matrix.

Principal components analysis of an H2O data frame using the power method to calculate the singular value decomposition of the Gram matrix.

h2o.prcomp(training_frame, x, model_id = NULL, validation_frame = NULL,
  ignore_const_cols = TRUE, score_each_iteration = FALSE,
  transform = c("NONE", "STANDARDIZE", "NORMALIZE", "DEMEAN", "DESCALE"),
  pca_method = c("GramSVD", "Power", "Randomized", "GLRM"),
  pca_implementation = c("MTJ_EVD_DENSEMATRIX", "MTJ_EVD_SYMMMATRIX",
  "MTJ_SVD_DENSEMATRIX", "JAMA"), k = 1, max_iterations = 1000,
  use_all_factor_levels = FALSE, compute_metrics = TRUE,
  impute_missing = FALSE, seed = -1, max_runtime_secs = 0)

Arguments

training_frame
Id of the training data frame.

x
A vector containing the character names of the predictors in the model.

model_id
Destination id for this model; auto-generated if not specified.

validation_frame
Id of the validation data frame.

ignore_const_cols
Logical. Ignore constant columns. Defaults to TRUE.

score_each_iteration
Logical. Whether to score during each iteration of model training. Defaults to FALSE.

transform
Transformation of training data Must be one of: "NONE", "STANDARDIZE", "NORMALIZE", "DEMEAN", "DESCALE". Defaults to NONE.

pca_method
Method for computing PCA (Caution: GLRM is currently experimental and unstable) Must be one of: "GramSVD", "Power", "Randomized", "GLRM". Defaults to GramSVD.

pca_implementation
Implementation for computing PCA (via SVD or EVD) Must be one of: "MTJ_EVD_DENSEMATRIX", "MTJ_EVD_SYMMMATRIX", "MTJ_SVD_DENSEMATRIX", "JAMA". Defaults to MTJ_EVD_SYMMMATRIX.

k
Rank of matrix approximation Defaults to 1.

max_iterations
Maximum training iterations Defaults to 1000.

use_all_factor_levels
Logical. Whether first factor level is included in each categorical expansion Defaults to FALSE.

compute_metrics
Logical. Whether to compute metrics on the training data Defaults to TRUE.

impute_missing
Logical. Whether to impute missing entries with the column mean Defaults to FALSE.

seed
Seed for random numbers (affects certain parts of the algo that are stochastic and those might or might not be enabled by default) Defaults to -1 (time-based random number).

max_runtime_secs
Maximum allowed runtime in seconds for model training. Use 0 to disable. Defaults to 0.

Value

Returns an object of class H2ODimReductionModel.

References

N. Halko, P.G. Martinsson, J.A. Tropp. Finding structure with randomness: Probabilistic algorithms for constructing approximate matrix decompositions[http://arxiv.org/abs/0909.4061]. SIAM Rev., Survey and Review section, Vol. 53, num. 2, pp. 217-288, June 2011.

See also

h2o.svd, h2o.glrm

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   15.82 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

ausPath <- system.file("extdata", "australia.csv", package="h2o")
australia.hex <- h2o.uploadFile(path = ausPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

h2o.prcomp(training_frame = australia.hex, k = 8, transform = "STANDARDIZE")

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Warning: _train: Dataset used may contain fewer number of rows due to removal of rows with NA/missing values.  If this is not desirable, set impute_missing argument in pca call to TRUE/True/true/... depending on the client language.

#> Model Details:
#> ==============
#> 
#> H2ODimReductionModel: pca
#> Model ID:  PCA_model_R_1512747612023_1459 
#> Importance of components: 
#>                             pc1      pc2      pc3      pc4      pc5      pc6
#> Standard deviation     1.750703 1.512142 1.031181 0.828313 0.608379 0.548136
#> Proportion of Variance 0.383120 0.285822 0.132917 0.085763 0.046266 0.037557
#> Cumulative Proportion  0.383120 0.668942 0.801859 0.887622 0.933887 0.971444
#>                             pc7      pc8
#> Standard deviation     0.418162 0.231495
#> Proportion of Variance 0.021857 0.006699
#> Cumulative Proportion  0.993301 1.000000
#> 
#> 
#> H2ODimReductionMetrics: pca
#> 
#> No model metrics available for PCA
#> 
#>