Set Levels of H2O Factor Column

Works on a single categorical vector. New domains must be aligned with the old domains. This call has SIDE EFFECTS and mutates the column in place (change of the levels will also affect all the frames that are referencing this column). If you want to make a copy of the column instead, use parameter in.place = FALSE.

h2o.setLevels(x, levels, in.place = TRUE)

Arguments

x
A single categorical column.

levels
A character vector specifying the new levels. The number of new levels must match the number of old levels.

in.place
Indicates whether new domain will be directly applied to the column (in place change) or if a copy of the column will be created with the given domain levels.

Examples

h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

iris.hex <- as.h2o(iris)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

new.levels <- c("setosa", "versicolor", "caroliniana")
iris.hex$Species <- h2o.setLevels(iris.hex$Species, new.levels, in.place = FALSE)

#> Error in is.H2OFrame(fr): object 'iris.hex' not found

h2o.levels(iris.hex$Species)

#> Error in is.H2OFrame(fr): object 'iris.hex' not found