Shut Down H2O Instance

Shut down the specified instance. All data will be lost.

h2o.shutdown(prompt = TRUE)

Arguments

prompt
A logical value indicating whether to prompt the user before shutting down the H2O server.

Details

This method checks if H2O is running at the specified IP address and port, and if it is, shuts down that H2O instance.

Note

Users must call h2o.shutdown explicitly in order to shut down the local H2O instance started by R. If R is closed before H2O, then an attempt will be made to automatically shut down H2O. This only applies to local instances started with h2o.init, not remote H2O servers.

WARNING

All data, models, and other values stored on the server will be lost! Only call this function if you and all other clients connected to the H2O server are finished and have saved your work.

See also

h2o.init

Examples

# Don't run automatically to prevent accidentally shutting down a cloud

# NOT RUN {
library(h2o)
h2o.init()
h2o.shutdown()
# }