Split an H2O Data Set

Split an existing H2O data set according to user-specified ratios. The number of subsets is always 1 more than the number of given ratios. Note that this does not give an exact split. H2O is designed to be efficient on big data using a probabilistic splitting method rather than an exact split. For example, when specifying a split of 0.75/0.25, H2O will produce a test/train split with an expected value of 0.75/0.25 rather than exactly 0.75/0.25. On small datasets, the sizes of the resulting splits will deviate from the expected value more than on big data, where they will be very close to exact.

h2o.splitFrame(data, ratios = 0.75, destination_frames, seed = -1)

Arguments

data
An H2OFrame object representing the dataste to split.

ratios
A numeric value or array indicating the ratio of total rows contained in each split. Must total up to less than 1.

destination_frames
An array of frame IDs equal to the number of ratios specified plus one.

seed
Random seed.

Value

Returns a list of split H2OFrame's

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 days 4 hours 
#>     H2O cluster timezone:       
#>     H2O data parsing timezone:  
#>     H2O cluster version:        3.17.0.99999 
#>     H2O cluster version age:    3 days  
#>     H2O cluster name:           terrytangyuan 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   17.59 GB 
#>     H2O cluster total cores:    8 
#>     H2O cluster allowed cores:  8 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         XGBoost, Algos, AutoML, Core V3, Core V4 
#>     R Version:                  R version 3.4.2 (2017-09-28) 
#> 

irisPath <- system.file("extdata", "iris.csv", package = "h2o")
iris.hex <- h2o.importFile(path = irisPath)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

iris.split <- h2o.splitFrame(iris.hex, ratios = c(0.2, 0.5))
head(iris.split[[1]])

#>    C1  C2  C3  C4          C5
#> 1 4.6 3.4 1.4 0.3 Iris-setosa
#> 2 5.4 3.7 1.5 0.2 Iris-setosa
#> 3 4.6 3.6 1.0 0.2 Iris-setosa
#> 4 4.8 3.1 1.6 0.2 Iris-setosa
#> 5 5.5 3.5 1.3 0.2 Iris-setosa
#> 6 5.0 3.5 1.3 0.3 Iris-setosa

summary(iris.split[[1]])

#> Warning: Approximated quantiles computed! If you are interested in exact quantiles, please pass the `exact_quantiles=TRUE` parameter.

#>  C1              C2              C3              C4              
#>  Min.   :4.600   Min.   :2.200   Min.   :1.000   Min.   :0.2000  
#>  1st Qu.:5.197   1st Qu.:2.798   1st Qu.:1.599   1st Qu.:0.2987  
#>  Median :5.600   Median :3.000   Median :4.100   Median :1.2983  
#>  Mean   :5.832   Mean   :3.040   Mean   :3.696   Mean   :1.1280  
#>  3rd Qu.:6.197   3rd Qu.:3.400   3rd Qu.:4.699   3rd Qu.:1.4999  
#>  Max.   :7.900   Max.   :3.800   Max.   :6.400   Max.   :2.3000  
#>  C5                 
#>  Iris-versicolor:14 
#>  Iris-setosa    : 7 
#>  Iris-virginica : 4 
#>                     
#>                     
#>