Plot an H2O Tabulate Heatmap

Plots the simple co-occurrence based tabulation of X vs Y as a heatmap, where X and Y are two Vecs in a given dataset. This function requires suggested ggplot2 package.

# S3 method for H2OTabulate
plot(x, xlab = x$cols[1], ylab = x$cols[2],
  base_size = 12, ...)

Arguments

x
An H2OTabulate object for which the heatmap plot is desired.

xlab
A title for the x-axis. Defaults to what is specified in the given H2OTabulate object.

ylab
A title for the y-axis. Defaults to what is specified in the given H2OTabulate object.

base_size
Base font size for plot.

...
additional arguments to pass on.

Value

Returns a ggplot2-based heatmap of co-occurance.

See also

link{h2o.tabulate}

Examples

library(h2o)
h2o.init()

#> Reading in config file: ./../../../../.h2oconfig
#> 
#> H2O is not running yet, starting it now...

#> Warning: cannot open file '/Users/terrytangyuan/h2o-3/h2o-r/h2o/branch.txt': No such file or directory

#> Error in file(con, "r"): cannot open the connection

df <- as.h2o(iris)

#> Error in h2o.getConnection(): No active connection to an H2O cluster. Did you run `h2o.init()` ?

tab <- h2o.tabulate(data = df, x = "Sepal.Length", y = "Petal.Width",
             weights_column = NULL, nbins_x = 10, nbins_y = 10)

#> Error in chk.H2OFrame(data): must be an H2OFrame

plot(tab)

#> Error in plot(tab): object 'tab' not found